% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadMIontology.R
\name{browseMIontology}
\alias{browseMIontology}
\title{browse Molecular Interaction ontology}
\usage{
browseMIontology()
}
\description{
browse Molecular Interaction ontology
}
\author{
Vitalii Kleshchevnikov
}
