% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centerDomains.R
\name{centerDomains}
\alias{centerDomains}
\title{Center QSLIMFinder datasets (InteractionSubsetFASTA_list) at domains}
\usage{
centerDomains(interactionFASTA_list, domain_res,
  domain_col = "IDs_domain_human", query_col = "IDs_interactor_viral",
  seed_col = "IDs_interactor_human")
}
\arguments{
\item{interactionFASTA_list}{object of class InteractionSubsetFASTA_list containing: FASTA sequences for interacting proteins, molecular interaction data they correspond to. Each element of a list contains input for individual QSLIMFinder run.}

\item{domain_res}{domain enrichment results (object of S3 class "XYZinteration_XZEmpiricalPval" (list), the output of permutationPval() call), or data.table with domain_col, query_col and seed_col}

\item{domain_col}{name of the column in \code{domain_res} containing domain ID, not relevant if \code{domain_res} is XYZinteration_XZEmpiricalPval}

\item{query_col}{name of the column in \code{domain_res} containing query protein ID, not relevant if \code{domain_res} is XYZinteration_XZEmpiricalPval}

\item{seed_col}{name of the column in \code{domain_res} containing seed protein ID, not relevant if \code{domain_res} is XYZinteration_XZEmpiricalPval}
}
\value{
object of class InteractionSubsetFASTA_list containing domain-centered datasets
}
\description{
Center QSLIMFinder datasets (InteractionSubsetFASTA_list) at domains
}
\examples{
forSLIMFinder = listInteractionSubsetFASTA(interaction_set1 = all_human_interaction,
                 interaction_set2 = all_viral_interaction,
                 seed_id_vect = proteins_w_signif_domains,
                 fasta = all.fasta,
                 single_interact_from_set2 = T, set1_only = T)
}
\seealso{
\code{\link{subset2setsBy1ID}}, \code{\link{recodeFASTA}}, \code{\link{interactionSubsetMapID}}, \code{\link{recodeANDinteractionSubsetFASTA}}, \code{\link{singleInteractFromSet2}}, \code{\link{listSingleInteractFromSet2}}, \code{\link{filterInteractionSubsetFASTA_list}}
}
\author{
Vitalii Kleshchevnikov
}
