% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanMITAB25.R
\name{cleanMITAB25}
\alias{cleanMITAB25}
\title{\code{cleanMITAB} extracts and cleans a set of columns of MITAB2.5}
\usage{
cleanMITAB25(mitab)
}
\arguments{
\item{mitab}{data.table containing molecular interaction data in MITAB 2.5 format}
}
\description{
function called internally by \code{\link{cleanMITAB}} if the format is MITAB2.5. No column names.
}
\author{
Vitalii Kleshchevnikov
}
