% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domainEnrichment.R
\name{domainEnrichment}
\alias{domainEnrichment}
\title{domainEnrichment: filter domains by background domain count => foldEnrichmentDist => foldEnrichmentPval => pval correction => plot pvals}
\usage{
domainEnrichment(backgr_domain_count, net, protein_annot, data, N = 1000,
  cores = NULL, seed = 1, all.data)
}
\description{
domainEnrichment: filter domains by background domain count => foldEnrichmentDist => foldEnrichmentPval => pval correction => plot pvals
}
