% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domainProteinPairMatch.R
\name{domainProteinPairMatch}
\alias{domainProteinPairMatch}
\title{Check if the protein in the set2 is associated with domain that is contained in the protein in the name and remove if not}
\usage{
domainProteinPairMatch(InteractionSubsetFASTA_list, domain_res,
  remove = T)
}
\arguments{
\item{InteractionSubsetFASTA_list}{object of class InteractionSubsetFASTA_list created by \code{\link{listInteractionSubsetFASTA}}}

\item{domain_res}{the result of domain enrichment analysis: object class XYZinteration_XZEmpiricalPval, the output of \code{\link{permutationPval}}. The domain should be in nodeZ, the set2 protein should be in the nodeX, the set1 protein should be in the nodeY.}

\item{remove}{logical, remove and return InteractionSubsetFASTA_list or just show which sequence and interaction sets have a protein in the set2 associated with domain that is contained in the protein in the name.}
}
\value{
object of class InteractionSubsetFASTA_list filtered for matching protein and containing: FASTA sequences for interacting proteins, molecular interaction data they correspond to. Each element of a list contains input for individual QSLIMFinder run.
}
\description{
Check if the protein in the set2 is associated with domain that is contained in the protein in the name and remove if not
}
\examples{
forSLIMFinder_Ready = domainProteinPairMatch(forSLIMFinder_Ready, domain_res, remove = T)
}
\seealso{
\code{\link{listInteractionSubsetFASTA}}, \code{\link{permutationPval}}
}
\author{
Vitalii Kleshchevnikov
}
