% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_fasta.R
\name{download_fasta}
\alias{download_fasta}
\title{Download FASTA for a list of canonical (P04637) and isoform (P04637-2) UniprotAC}
\usage{
download_fasta(uniprot_ac, file_name)
}
\arguments{
\item{uniprot_ac}{character vector containg canonical and isoform UniprotAC}

\item{file_name}{file name and directory to save the result ("./dir/filename")}
}
\value{
nothing
}
\description{
Download FASTA for a list of canonical (P04637) and isoform (P04637-2) UniprotAC
}
\details{
\code{download_fasta} is intended to download from Uniprot and write fasta sequence file (renaming the name of the sequence given by UniProt to UniprotAC only) given generic (P04637) or isoform (P04637-2) UniprotAC. If the file already contains the sequences for given UniprotAC you get a message, if some sequences are missing all sequences will be reloaded.
}
\examples{
download_fasta(uniprot_ac = c("P04637", "P04637-2"), file_name = "my_canonical_and_isoform.fasta")
}
\seealso{
\code{\link{download_fasta_postproc}}
}
\author{
Vitalii Kleshchevnikov
}
