% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadFasta.R
\name{downloadFasta}
\alias{downloadFasta}
\alias{loadAllFASTA}
\title{Download FASTA for a list of canonical (P04637) and isoform (P04637-2) UniprotAC}
\usage{
downloadFasta(uniprot_ac, file_name)


  loadAllFASTA(uniprot_url = "ftp://ftp.uniprot.org/pub/databases/uniprot/current_release/knowledgebase/complete/uniprot_sprot.fasta.gz",
  uniprot_ac, file = "sp.fasta.gz")
}
\arguments{
\item{uniprot_ac}{character vector containg canonical and isoform UniprotAC}

\item{file_name}{file name and directory to save the result ("./dir/filename")}

\item{uniprot_url}{file name and directory to save the result ("./dir/filename")}

\item{uniprot_ac}{character vector containg canonical and isoform UniprotAC}
}
\value{
nothing

AAStringSet
}
\description{
Download FASTA for a list of canonical (P04637) and isoform (P04637-2) UniprotAC

Load all FASTA from single UniProt file and filter by uniprot_ac
}
\details{
\code{downloadFasta} is intended to download from Uniprot and write fasta sequence file (renaming the name of the sequence given by UniProt to UniprotAC only) given generic (P04637) or isoform (P04637-2) UniprotAC. If the file already contains the sequences for given UniprotAC you get a message, if some sequences are missing all sequences will be reloaded.
}
\examples{
downloadFasta(uniprot_ac = c("P04637", "P04637-2"), file_name = "my_canonical_and_isoform.fasta")
}
\seealso{
\code{\link{downloadFasta_postproc}}
}
\author{
Vitalii Kleshchevnikov

Vitalii Kleshchevnikov
}
