% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadFastaPostproc.R
\name{downloadFastaPostproc}
\alias{downloadFastaPostproc}
\title{Download FASTA for a list of post-processed chain ID (e.g. PRO_0000261216)}
\usage{
downloadFastaPostproc(postproc_id, file_name)
}
\arguments{
\item{file_name}{file name and directory to save the result ("./dir/filename")}

\item{uniprot_ac}{character vector containg post-processed chain ID}
}
\value{
nothing
}
\description{
Download FASTA for a list of post-processed chain ID (e.g. PRO_0000261216)
}
\details{
function to download from Uniprot and write fasta sequence files (renaming the UniProt-given name of the sequence to proteinID-FeatureID: P04591-PRO_0000261216) given post-processed chain ID, such as PRO_0000261216. If the file already contains the sequences for given post-processed chain ID you get a message, if some sequences are missing all sequences will be reloaded.
}
\examples{
downloadFastaPostproc(postproc_id = c("PRO_0000261216"), file_name = "my_post_processed_chain.fasta")
}
\seealso{
\code{\link{download_fasta}}
}
\author{
Vitalii Kleshchevnikov
}
