% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgelist2degreeFreq.R
\name{edgelist2degreeFreq}
\alias{edgelist2degreeFreq}
\title{Find the degree distribution given edgelist (pair_id) and a list of proteins which degrees matter}
\usage{
edgelist2degreeFreq(mitab, prots = NULL, sep = "\\\\|")
}
\arguments{
\item{mitab}{data.table containing molecular interaction data, including pair_id}

\item{prots}{list of relevant proteins, degree distribution is based on the prevalence of specific degrees in this list, all proteins in \code{mitab} by default}

\item{sep}{ID separator in pair_id, passed to \code{\link{edgelist2degree}}}
}
\description{
Find the degree distribution given edgelist (pair_id) and a list of proteins which degrees matter
}
\details{
unique pair_id is the unique identifier of the undirected interaction: ordered alphabetically and concatenated interacting molecule IDs \code{\link{cleanMITAB}}
}
\examples{
# download full human interactome (clean = TRUE is necessary to produce the right input for edgelist2degree)
full = fullInteractome(taxid = "9606", database = "imex", format = "tab25", clean = TRUE, protein_only = TRUE)
degree_freq = edgelist2degreeFreq(full)
plot(degree_freq$N, degree_freq$degree_freq,
     log = "xy",
     main = "degree distribution \\n - human-human protein interactions -",
     ylab = "degree frequency, log-scale",
     xlab = "degree, log-scale",
     las=1)
}
\author{
Vitalii Kleshchevnikov
}
