% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SLIMFinderOcc2GRanges.R
\name{ELMdb2GRanges}
\alias{ELMdb2GRanges}
\title{Download and read ELM database motif occurrences}
\usage{
ELMdb2GRanges(dbfile = "../viral_project/data_files/instances9606.gff",
  dburl = "http://elm.eu.org/instances.gff?q=None&taxon=Homo\%20sapiens&instance_logic=",
  tsvurl = "http://elm.eu.org/instances.tsv?q=None&taxon=Homo\%20sapiens&instance_logic=",
  tsvfile = "../viral_project/data_files/instances9606.tsv")
}
\arguments{
\item{dbfile}{a path to a gff (txt) file containing ELM database motif occurrences}

\item{dburl}{url where to get ELM database containing motif occurrences}
}
\value{
Genomic Ranges object containing ELM database motif occurrences
}
\description{
Download and read ELM database motif occurrences
}
\details{
ELM database containing motif occurrences in human proteins (9606): "http://elm.eu.org/instances.gff?q=None&taxon=Homo%20sapiens&instance_logic="

ELM database containing motif occurrences in viral proteins (10239): "http://elm.eu.org/instances.gff?q=all&taxon=irus&instance_logic="
}
\examples{
instances9606 = ELMdb2GRanges(dbfile = "../viral_project/data_files/instances9606.gff",
                  dburl = "http://elm.eu.org/instances.gff?q=None&taxon=Homo\%20sapiens&instance_logic=")
instances10239 = ELMdb2GRanges(dbfile = "../viral_project/data_files/instances10239.gff",
                  dburl = "http://elm.eu.org/instances.gff?q=all&taxon=irus&instance_logic=")
}
\seealso{
\code{\link{SLIMFinderOcc2GRanges}}, \code{\link{GRangesINinteractionSubsetFASTA}}
}
\author{
Vitalii Kleshchevnikov
}
