% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractInteractors.R
\name{extractInteractors}
\alias{extractInteractors}
\alias{uniqueNinteractions}
\alias{uniqueNinteractors}
\title{Subset MITAB to retrieve interactors, the number of interactions ot interactors}
\usage{
extractInteractors(cleanMITAB, taxid = NULL, inverse_filter = F)

uniqueNinteractions(cleanMITAB)

uniqueNinteractors(cleanMITAB, taxid = NULL, inverse_filter = F)
}
\arguments{
\item{cleanMITAB}{object of any clean_MItab class (the class that contains "clean_MItab" in it's name and is initially produced by  \code{\link[PItools]{cleanMITAB}})}

\item{taxid}{filter \code{cleanMITAB} using this list of taxonomy IDs}

\item{inverse_filter}{logical, inverse filtering criteria}
}
\value{
character vector of interactor IDs or an integer count of interactions or interactors
}
\description{
subset molecular interaction data (cleaned MITAB format in a data.table object) with a list of interactors
}
\author{
Vitalii Kleshchevnikov
}
