% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SLIMFinderOcc2GRanges.R
\name{filterBYmotifType}
\alias{filterBYmotifType}
\title{Filter ELM database motif occurrence GRanges by motif type}
\usage{
filterBYmotifType(instances, motif_types = c("DOC", "MOD", "LIG", "DEG",
  "CLV", "TRG"))
}
\arguments{
\item{instances}{ELM database motif occurrence GRanges object returned by \code{\link{ELMdb2GRanges}}}

\item{motif_types}{character vector of motif types}
}
\value{
Genomic Ranges object filtered by motif type (ELMType column)
}
\description{
Filter ELM database motif occurrence GRanges by motif type
}
\examples{
instances9606 = ELMdb2GRanges(dbfile = "../viral_project/data_files/instances9606.gff",
                  dburl = "http://elm.eu.org/instances.gff?q=None&taxon=Homo\%20sapiens&instance_logic=")
instances9606 = filterBYmotifType(instances9606, motif_types = c("DOC", "MOD", "LIG", "DEG", "CLV", "TRG"))
}
\seealso{
\code{\link{SLIMFinderOcc2GRanges}}, \code{\link{GRangesINinteractionSubsetFASTA}}
}
\author{
Vitalii Kleshchevnikov
}
