% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterInteractionSubsetFASTA_list.R
\name{filterInteractionSubsetFASTA_list}
\alias{filterInteractionSubsetFASTA_list}
\title{Filter InteractionSubsetFASTA_list by the size of the interaction sets (number of proteins in each)}
\usage{
filterInteractionSubsetFASTA_list(interactionFASTA_list,
  length_set1_min = 0, length_set2_min = 0, length_set1_max = Inf,
  length_set2_max = Inf)
}
\arguments{
\item{interactionFASTA_list}{object of class InteractionSubsetFASTA_list created by \code{\link{listInteractionSubsetFASTA}}}

\item{length_set1_min}{mininal length of set 1, for example, if a set has less than 1 protein QSLIMFinder wouldn't work.}

\item{length_set2_min}{mininal length of set 2}

\item{length_set1_max}{maximal length of set 1, for example, QSLIMFinder has an upper limit of 500 sequences}

\item{length_set2_max}{maximal length of set 2}
}
\value{
object of class InteractionSubsetFASTA_list containing: FASTA sequences for interacting proteins, molecular interaction data they correspond to. Each element of a list contains input for individual QSLIMFinder run.
}
\description{
Filter InteractionSubsetFASTA_list by the size of the interaction sets (number of proteins in each)
}
\examples{
forSLIMFinder_Ready = filterInteractionSubsetFASTA_list(forSLIMFinder,  length_set1_min = 3, length_set2_min = 1)
}
\seealso{
\code{\link{listInteractionSubsetFASTA}}
}
\author{
Vitalii Kleshchevnikov
}
