% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findOverlapsBench.R
\name{findOverlapsBench}
\alias{findOverlapsBench}
\title{Find occuring GRanges overlapping with the benchmarking set (positive + negative examples)}
\usage{
findOverlapsBench(occuring, benchmarking, predictor_col = "Sig",
  labels_col = "for_benchmarking", normalise = T,
  select_predictor_per_range = min, ...)
}
\arguments{
\item{occuring}{GRanges object, occuring feature instances}

\item{benchmarking}{GRanges object, feature instances for benchmarking}

\item{predictor_col}{name of the metadata column in \code{occuring} that contains 1 - predictor value (such as p-value)}

\item{labels_col}{name of the metadata column in \code{benchmarking} that contains labels for benchmarking (for, example 0 for negative and 1 for positive)}

\item{normalise}{logical, normalise predictor value, just in case predictor doesn't span the full range between 0 ... 1}

\item{...}{arguments such as \code{maxgap}, \code{minoverlap} passed to \code{\link[GenomicRanges]{findOverlaps}}}
}
\value{
Genomic Ranges list object containing \code{N} randomised GRanges objects
}
\description{
Find occuring GRanges overlapping with the benchmarking set (positive + negative examples)
}
\seealso{
\code{\link{ELMdb2GRanges}}, \code{\link{GRangesINinteractionSubsetFASTA}}
}
\author{
Vitalii Kleshchevnikov
}
