% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foldEnrichment.R
\name{foldEnrichment}
\alias{foldEnrichment}
\title{\code{foldEnrichment} calculates fold enrichment of a domain (or other feature) among interacting partners of a protein}
\usage{
foldEnrichment(net, protein_annot, frequency = T)
}
\arguments{
\item{net}{data.table that specifies the PPI network:
1. first column specifies a protein ("IDs_interactor_viral")
2. second column specifies it's interacting partner which annotations are of interest ("IDs_interactor_human")
3. third column specifies the degree of a protein in the first column ("IDs_interactor_viral_degree")}

\item{protein_annot}{data.table that specifies annotations for a protein in the second column of \code{net}
1. first column specifies a protein ("IDs_interactor_human")
2. second column specifies features of that protein ("IDs_domain_human")
3. third column specifies background frequency of those features ("domain_frequency)}

\item{frequency}{fold enrichment or frequency in a set (if TRUE - frequency)}
}
\value{
data.table containing fold enrichment for each domain - protein pair
}
\description{
\code{foldEnrichment} calculates fold enrichment of a domain (or other feature) among interacting partners of a protein
}
\author{
Vitalii Kleshchevnikov
}
