% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foldEnrichmentDist.R
\name{foldEnrichmentDist}
\alias{foldEnrichmentDist}
\title{\code{foldEnrichmentDist} finds fold_enrichment distribution (in any domain or other feature) for each protein}
\usage{
foldEnrichmentDist(net, protein_annot, N = 1000, cores = NULL)
}
\arguments{
\item{net}{data.table that specifies the PPI network:
1. first column specifies a protein ("IDs_interactor_viral")
2. second column specifies it's interacting partner which annotations are of interest ("IDs_interactor_human")
3. third column specifies the degree of a protein in the first column ("IDs_interactor_viral_degree")}

\item{protein_annot}{data.table that specifies annotations for a protein in the second column of \code{net}
1. first column specifies a protein ("IDs_interactor_human")
2. second column specifies features of that protein ("IDs_domain_human")
3. third column specifies background frequency of those features ("domain_frequency)}
}
\value{
data.table containing fold enrichment for each domain - protein pair
}
\description{
\code{foldEnrichmentDist} finds fold_enrichment distribution (in any domain or other feature) for each protein in the first column of net. For more details: \code{\link{foldEnrichment}}
}
\author{
Vitalii Kleshchevnikov
}
