% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullInteractome.R
\name{fullInteractome}
\alias{fullInteractome}
\title{Retrieve interactome (proteins only or not) of a given taxonomic species}
\usage{
fullInteractome(MITABdata = NULL, taxid = 9606, database = "imex",
  format = "tab25", clean = TRUE, protein_only = TRUE, directory = NULL,
  releaseORdate = NULL, remove_obsolete_id = F)
}
\arguments{
\item{MITABdata}{object of class "RAW_MItab25" or "RAW_MItab27" (list) containing molecular interaction data as returned by \code{\link{queryPSICQUICrlib}}, default in NULL}

\item{taxid}{character (1L), taxonomy id of the species which interaction participants should belong to, default is "9606" (which is human)}

\item{database}{character (1L), argument for \code{\link{queryPSICQUIC}}, default is "imex" alternative to which is "IntActFTP"}

\item{format}{character (1L), argument for \code{\link{queryPSICQUIC}}, default is "tab25"}

\item{clean}{logical (1L), if TRUE extract specific information using \code{\link{cleanMITAB}}, default is TRUE}

\item{protein_only}{logical (1L), if TRUE the interaction participants are restricted to proteins (exclude other types of molecules such as RNA or small molecules), default is TRUE}

\item{directory}{directory where to store the data, if NULL the data is stored in <R-package-library>/MItools/data}

\item{releaseORdate}{character, if data has already been downloaded: which IntAct release or download date to read}

\item{remove_obsolete_id}{logical (1L), remove interactions in which one of the partners is encoded as obsolete UniProtKB accession (ID)}
}
\value{
object of class `input class`_fullInteractome containing data.table containing molecular interaction data in either of these two formats:

if \code{clean} is TRUE: contains columns as described in \code{\link{cleanMITAB}};

if \code{clean} is FALSE: contains a standard set of columns for MITAB2.5 or MITAB2.7 depending on \code{format};
}
\description{
Retrieve interactome (proteins only or not) of a given taxonomic species from a specified database. Interactome can be additionally cleaned and includes only specific information: \code{\link{cleanMITAB}}
}
\details{
\code{taxid} is used to query specified database using PSICQUIC client, only interactions in which both participants belong the \code{taxid} are retured (\code{"taxidA:9606 AND taxidB:9606"}, not \code{"species:9606"}). Details: \code{\link{queryPSICQUIC}}

\code{fullInteractome} can be used to retrive interactome data using PSICQUIC service using \code{\link{queryPSICQUIC}}, clean and select specific columns using \code{\link{cleanMITAB}} and filter resulting dataset for protein-protein interaction only. This is the default option.

Alternatively, \code{fullInteractome} can only retrive interactome data using PSICQUIC service without cleaning of filtering.

Another option is to supply \code{MITABdata} to be cleaned and filtered

Finally, you can avoid using PSICQUIC service and download data from IntAct ftp by selecting database argument "IntActFTP". This is much faster but larger requires larger download (>3Gb) and is more computationally intensive for processing. As of 7.09.2017 "IntActFTP" provides access to DIP data, while "imex" doesn't.
}
\examples{
# retrive a full set of human (9606) protein-protein interactions from IMEx databases in MITAB2.5 format, clean and select specific columns
full = fullInteractome(taxid = "9606", database = "imex", format = "tab25", clean = TRUE, protein_only = TRUE)
# Do the same using IntAct ftp instead of the PSICQUIC webservice, except that, IntAct ftp option always outputs tab27 format
full = fullInteractome(taxid = "9606", database = "IntActFTP", clean = TRUE, protein_only = TRUE)

# retrive a full set of human (9606) protein-protein interactions from IMEx databases in MITAB2.5 format, clean and select specific columns; save it to the specific directory inside working directory
full = fullInteractome(taxid = "9606", database = "imex", format = "tab25", clean = TRUE, protein_only = TRUE, directory = "./data/")
}
\seealso{
\code{\link{interSpeciesInteractome}}
}
\author{
Vitalii Kleshchevnikov
}
