% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRangesINinteractionSubsetFASTA.R
\name{GRangesINinteractionSubsetFASTA}
\alias{GRangesINinteractionSubsetFASTA}
\title{Check if proteins in GRanges were used for motif search using QSLIMFinder (SLIMFinder)}
\usage{
GRangesINinteractionSubsetFASTA(grange, interactionSubsetFASTA,
  query_only = F, clustermq = F)
}
\arguments{
\item{interactionSubsetFASTA}{object of class InteractionSubsetFASTA_list containing: FASTA sequences for interacting proteins, molecular interaction data they correspond to. Each element of a list contains input for individual QSLIMFinder run.}

\item{query_only}{return only GRanges for query proteins}

\item{GRanges}{object containing instances of linear motifs or domains in proteins (identified by UniProtKB accession)}
}
\value{
list containing 2 vectors: which GRanges are in InteractionSubsetFASTA_list and seqlengths vector for these GRanges
}
\description{
Check if proteins in GRanges were used for motif search using QSLIMFinder (SLIMFinder)
}
\seealso{
\code{\link{ELMdb2GRanges}}, \code{\link{runQSLIMFinder}}, \code{\link{SLIMFinderOcc2GRanges}}, \code{\link{listInteractionSubsetFASTA}}, \code{\link{randomGRanges}}
}
\author{
Vitalii Kleshchevnikov
}
