% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humanViralDegree.R
\name{humanViralDegree}
\alias{humanViralDegree}
\title{Compare viral and human protein degree distribution, filter by interaction detection method or PMIDs}
\usage{
humanViralDegree(data = NULL, directory = "./data_files/",
  Interaction_detection_methods = NULL, Identification_method = NULL,
  PMIDs = NULL, inverse_filter = F, data_name = "")
}
\arguments{
\item{data}{a list contatining PPI data as generated by \code{\link[PItools]{loadHumanViralPPI}}, if NULL, data will be downloaded internally}

\item{directory}{where to find / keep PPI data (arg for \code{\link[PItools]{fullInteractome}}, \code{\link[PItools]{interSpeciesInteractome}})) and MI ontology (arg for \code{\link[PItools]{subsetMITABbyMethod}})}

\item{Interaction_detection_methods}{arg for \code{\link[PItools]{subsetMITABbyMethod}}}

\item{Identification_method}{arg for \code{\link[PItools]{subsetMITABbyMethod}}}

\item{PMIDs}{arg for \code{\link[PItools]{subsetMITABbyPMIDs}}}

\item{inverse_filter}{logical, inverse filtering criteria}

\item{data_name}{to be displayed on plot}
}
\value{
data.table to be ggplotted to compare viral and human protein degree distributions
}
\description{
subset molecular interaction data (cleaned MITAB format in a data.table object) with a list of interactors
}
\examples{
{
# Load data. This can be done internally by humanViralDegree(),
# however, it takes time (minutes) so it's recommended to load data separately
# if it will be used multiple times
HumanViralPPI = loadHumanViralPPI(directory = "./data_files/")

# prepare all data for the degree distribution analysis
degree_distributions = humanViralDegree(data = HumanViralPPI,
  directory = "./data_files/", data_name = "Full IntAct")

# prepare two-hybrid data for the degree distribution analysis
degree_distributions = humanViralDegree(data = HumanViralPPI,
  directory = "./data_files/", Interaction_detection_methods = "MI:0018",
  data_name = "two-hybrid data")

# prepare AP-MS data for the degree distribution analysis
degree_distributions = humanViralDegree(data = HumanViralPPI,
  directory = "./data_files/", Interaction_detection_methods = "MI:0004",
  Identification_method = "MI:0433", PMIDs = NULL, inverse_filter = F,
  data_name = "AP-MS data")

# to see if viral-interacting human proteins are special in the human network,
# look at the Vidal published and unpublished datasets (only human-human network
# is modified (data for the top 2 plots))
degree_distributions = humanViralDegree(data = HumanViralPPI,
  directory = "./data_files/", PMIDs = c("25416956", "unassigned1304"),
  data_name = "Vidal published and unpublished")
# to see if viral-interacting human proteins are special in the human network,
# do the same for Matthias Mann 2015 paper dataset
degree_distributions = humanViralDegree(data = HumanViralPPI,
  directory = "./data_files/", PMIDs = "26496610",
  data_name = "Matthias Mann 2015 paper")
  }
}
\author{
Vitalii Kleshchevnikov
}
