% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactionSubsetFASTA.R
\name{interactionSubsetFASTA}
\alias{interactionSubsetFASTA}
\title{FASTA sequences of interactors of a specific protein}
\usage{
interactionSubsetFASTA(int_subset, fasta)
}
\arguments{
\item{int_subset}{interaction_subset_from_2_sets object renamed using interactionSubsetMapID}

\item{fasta}{AAStringSet renamed by recodeFASTA}
}
\value{
object of class interactionSubsetFASTA containing: FASTA sequences for interacting proteins, original interaction_subset_from_2_sets
}
\description{
FASTA sequences of interactors of a specific protein
}
\examples{
subset = subset2setsBy1ID(interaction_set1 = all_human_interaction,
                        interaction_set2 = all_viral_interaction,
                        seed_id = proteins_w_signif_domains[2])
# subset FASTA
subset_fasta = interactionSubsetFASTA(int_subset = subset, fasta = all.fasta$fasta)
# rename FASTA sequences to original names
subset_fasta = recodeFASTA(subset_fasta, all.fasta$names_mapping, i = 1)
}
\seealso{
\code{\link{subset2setsBy1ID}}, \code{\link{recodeFASTA}}, \code{\link{interactionSubsetMapID}}
}
\author{
Vitalii Kleshchevnikov
}
