% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset2setsBy1ID.R
\name{interactionSubsetMapID}
\alias{interactionSubsetMapID}
\title{interaction_subset_from_2_sets renames ids in the interaction_subset_from_2_sets object}
\usage{
interactionSubsetMapID(subset, mapping)
}
\arguments{
\item{subset}{interaction_subset_from_2_sets, the output of \code{\link{subset2setsBy1ID}}}

\item{mapping}{how to translate original IDs to new IDs}
}
\value{
object of class interaction_subset_from_2_sets (list) with ids_all slot containing new names and slot ids_all_old containing original names
}
\description{
interaction_subset_from_2_sets renames ids in the interaction_subset_from_2_sets object
}
\examples{
subset = interactionSubsetMapID(subset, all.fasta$names_mapping)
}
\seealso{
\code{\link{subset2setsBy1ID}}, \code{\link{recodeFASTA}}
}
\author{
Vitalii Kleshchevnikov
}
