% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactorFeatureTypes.R
\name{interactorFeatureTypes}
\alias{interactorFeatureTypes}
\title{discover types of interactor features in MITAB 2.7}
\usage{
interactorFeatureTypes(mitab)
}
\arguments{
\item{mitab}{data.table containing molecular interaction data in MITAB 2.5 or 2.7 formats. Details: \code{\link{queryPSICQUIC}}}
}
\description{
discover types of interactor features in MITAB 2.7
}
\details{
\code{interactorFeatureTypes} extracts types of interactor features from Features_interactor_A and Features_interactor_B in MITAB 2.7 format \link{https://psicquic.github.io/MITAB27Format.html}

All interactor features are decribed in the molecular interaction ontology (MI): \link{http://www.ebi.ac.uk/ols/ontologies/MI/terms?obo_id=MI:0116}
}
\seealso{
\code{\link{cleanMITAB}}
}
\author{
Vitalii Kleshchevnikov
}
