% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactors2interactions.R
\name{interactors2interactions}
\alias{interactors2interactions}
\title{Retrieve interactions for a list of interactors (given data table, not using webservice)}
\usage{
interactors2interactions(MITABdata, interactorsList)
}
\arguments{
\item{MITABdata}{data.table with the pair_id column - unique identifier of the undirected interaction: ordered alphabetically and concatenated interacting molecule IDs}

\item{interactorsList}{filter \code{MITABdata} using this list of interactors}
}
\value{
data.table, a subset of MITABdata that contains interactions for interactorsList molecules
}
\description{
subset molecular interaction data (cleaned MITAB format in a data.table object) with a list of interactors
}
\author{
Vitalii Kleshchevnikov
}
