% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lastIntActRelease.R
\name{lastIntActRelease}
\alias{lastIntActRelease}
\title{last IntAct (all IMEx databases) release}
\usage{
lastIntActRelease()
}
\value{
character string specifying the year, month, day, and the hour when the IntAct data file was last modified
}
\description{
last IntAct (all IMEx databases) release
}
