% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mInteractionSubsetFASTA.R
\name{listInteractionSubsetFASTA}
\alias{listInteractionSubsetFASTA}
\title{get FASTA sequences of proteins that interact with seed proteins (datasets for QSLIMFinder)}
\usage{
listInteractionSubsetFASTA(interaction_set1, interaction_set2,
  seed_id_vect, fasta, single_interact_from_set2 = T, set1_only = T)
}
\arguments{
\item{interaction_set1}{cleanMItab object containing molecular interactions (such as all interactions between human proteins)}

\item{interaction_set2}{a different cleanMItab object containing molecular interactions (conceptually different source of interaction data, such as other type of interactions or human-viral interactions)}

\item{seed_id_vect}{a vector of interactor IDs for which to retrieve interactions (for each separately)}

\item{fasta}{AAStringSet containing sequences for all proteins in interaction_set1 and interaction_set2}

\item{single_interact_from_set2}{logical, split sequence sets to contain only one protein from interaction_set2 (only one query protein for QSLIMFinder). If FALSE, set2 will contain all proteins that interact with an element of \code{seed_id_vect} (which means multiple query proteins for QSLIMFinder).}

\item{set1_only}{logical, only relevant if \code{single_interact_from_set2 = TRUE}, sequence set1 should contain only proteins that interact with an element of seed_id_vect in interaction_set1. If FALSE, proteins that interact with an element of seed_id_vect in interaction_set2 but are not a single query protein are also included. Argument for \code{\link{singleInteractFromSet2}}}
}
\value{
object of class InteractionSubsetFASTA_list containing: FASTA sequences for interacting proteins, molecular interaction data they correspond to. Each element of a list contains input for individual QSLIMFinder run.
}
\description{
get FASTA sequences of proteins that interact with seed proteins (datasets for QSLIMFinder)
}
\examples{
forSLIMFinder = listInteractionSubsetFASTA(interaction_set1 = all_human_interaction,
                 interaction_set2 = all_viral_interaction,
                 seed_id_vect = proteins_w_signif_domains,
                 fasta = all.fasta,
                 single_interact_from_set2 = T, set1_only = T)
}
\seealso{
\code{\link{subset2setsBy1ID}}, \code{\link{recodeFASTA}}, \code{\link{interactionSubsetMapID}}, \code{\link{recodeANDinteractionSubsetFASTA}}, \code{\link{singleInteractFromSet2}}, \code{\link{listSingleInteractFromSet2}}, \code{\link{filterInteractionSubsetFASTA_list}}
}
\author{
Vitalii Kleshchevnikov
}
