% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mInteractionSubsetFASTA.R
\name{listSingleInteractFromSet2}
\alias{listSingleInteractFromSet2}
\title{Split interaction subset from 2 sets so that each elements contains only one of the proteins from set 2, and retrieve FASTA sequences for each}
\usage{
listSingleInteractFromSet2(subset1, single_interact_from_set2, set1_only,
  fasta)
}
\arguments{
\item{subset1}{interaction_subset_from_2_sets object produced by \code{\link{subset2setsBy1ID}}}

\item{single_interact_from_set2}{logical, split sequence sets to contain only one protein from interaction_set2 (only one query protein for QSLIMFinder). If FALSE, set2 will contain all proteins that interact with an element of \code{seed_id_vect} (which means multiple query proteins for QSLIMFinder).}

\item{set1_only}{logical, only relevant if \code{single_interact_from_set2 = TRUE}, sequence set1 should contain only proteins that interact with an element of seed_id_vect in interaction_set1. If FALSE, proteins that interact with an element of seed_id_vect in interaction_set2 with the exception of the single one () are also included. Argument for \code{\link{singleInteractFromSet2}}}

\item{fasta}{list containing AAStringSet renamed by \code{\link{recodeFASTA}} and mapping table from \code{\link{recodeFASTA}}}
}
\value{
object of class InteractionSubsetFASTA_list containing: FASTA sequences for interacting proteins, molecular interaction data they correspond to. Each element of a list contains input for individual QSLIMFinder run.
}
\description{
Split interaction subset from 2 sets so that each elements contains only one of the proteins from set 2, and retrieve FASTA sequences for each
}
\examples{
subset1_fasta = listSingleInteractFromSet2(subset1, single_interact_from_set2, set1_only, fasta)
}
\seealso{
\code{\link{subset2setsBy1ID}}, \code{\link{listInteractionSubsetFASTA}}
}
\author{
Vitalii Kleshchevnikov
}
