% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadTaxIDProteome.R
\name{loadAllIDProteins}
\alias{loadAllIDProteins}
\title{load the list of all protein ID in UniProtKB}
\usage{
loadAllIDProteins(dir)
}
\arguments{
\item{dir}{directory where to save/look for the local copy}
}
\value{
\code{loadAllIDProteins} returns list containing the character vector of all UniProt accessions (saves it to a file)
}
\description{
loadAllIDProteins function loads all UniProt accessions for a given taxonomy ID
}
\author{
Vitalii Kleshchevnikov

Vitalii Kleshchevnikov
}
