% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadBioplex.R
\name{loadBioplex}
\alias{loadBioplex}
\title{load Bioplex data}
\usage{
loadBioplex(dir = "./",
  url = "http://bioplex.hms.harvard.edu/data/BioPlex_2.3_interactionList.tsv")
}
\arguments{
\item{dir}{directory where to save/look for the local copy}
}
\value{
\code{loadBioplex} saves loadBioplex data to a file, returns the object of class clean_MItab25
}
\description{
\code{loadBioplex} saves and/or loads the data stored on Bioplex website \link{http://bioplex.hms.harvard.edu/downloadInteractions.php}. Depends on the tsv file containing the following columns:
Gene A	Gene B	Uniprot A	Uniprot B	Symbol A	Symbol B	p(Wrong)	p(No Interaction)	p(Interaction)
}
\examples{
loadBioplex("./")
}
\author{
Vitalii Kleshchevnikov

Vitalii Kleshchevnikov
}
