% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humanViralDegree.R
\name{loadHumanViralPPI}
\alias{loadHumanViralPPI}
\title{load human-human, human-viral and viral-viral PPI data from IntActFTP}
\usage{
loadHumanViralPPI(directory = "./data_files/",
  loadIntActFTP_dir = "./data_files/IntActRelease_2017Nov13/",
  release = "2017Nov13")
}
\arguments{
\item{directory}{to find / keep PPI data (arg for \code{\link[PItools]{fullInteractome}}, \code{\link[PItools]{interSpeciesInteractome}}))}

\item{loadIntActFTP_dir}{where to find specific release (arg for \code{\link[PItools]{loadIntActFTP}})}

\item{release}{release date}
}
\description{
use \code{\link[PItools]{interSpeciesInteractome}} and \code{\link[PItools]{fullInteractome}} to retrieve and load human-human, human-viral and viral-viral PPI data from IntActFTP

to be used with \code{\link[PItools]{humanViralDegree}}
}
\author{
Vitalii Kleshchevnikov
}
