% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryIntActFTP.R
\name{loadIntActFTP}
\alias{loadIntActFTP}
\title{load all data from IntAct ftp}
\usage{
loadIntActFTP(dir)
}
\arguments{
\item{file}{directory where to save/look for the local copy}
}
\value{
saves intact.txt to a file, returns the object of class RAW_MItab27
}
\description{
loadIntActFTP function loads all data that is stored in \link{ftp://ftp.ebi.ac.uk/pub/databases/intact/current/psimitab/} in intact.txt file. Details: \link{http://www.ebi.ac.uk/intact/downloads}. This file contains the data from the following databases "IntAct", "MINT", "DIP", "bhf-ucl", "MPIDB", "MatrixDB", "HPIDb","I2D-IMEx","InnateDB-IMEx", "MolCon", "UniProt", "MBInfo"
}
\examples{
loadIntActFTP("./")
}
\author{
Vitalii Kleshchevnikov

Vitalii Kleshchevnikov
}
