% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadInterProTree.R
\name{loadInterProTree}
\alias{loadInterProTree}
\title{Load the hierarchy of relationships between InterPro's entries}
\usage{
loadInterProTree(filename)
}
\arguments{
\item{filename}{character, where to download and / or where to find InterPro Entry relationships tree file. Details: \link{https://www.ebi.ac.uk/interpro/download.html}}

\item{tree}{object of class "InterProTree", the output of \code{loadInterProTree} function}

\item{ParentChildTreeLines}{argument for non-exported \code{loadChildrenForLevel}: the lines of the InterPro Entry relationships tree file (character)}

\item{levels}{argument for non-exported \code{loadChildrenForLevel}: the mapping between the levels and the lines in ParentChildTreeLines}

\item{level}{argument for non-exported \code{loadChildrenForLevel}: integer, which level to get children for}

\item{regex}{argument for non-exported \code{loadChildrenForLevel}: regex that matches InterProID (IPR[[:digit:]]{6})}

\item{allchildren}{argument for non-exported \code{loadChildrenForLevel}: whether to return all children or only a specific level}
}
\value{
object of S3 class "InterProTree" containing data.table that provides the mapping between layers, the lines of the InterPro Entry relationships tree file (character) and the mapping between the levels and the lines in ParentChildTreeLines
}
\description{
Load the hierarchy of relationships between InterPro's entries
}
\examples{
InterProScan_domains_nonred = collapseByInterProID(InterProScan_features = InterProScan_domains, id_col = "Dbxref")
}
\seealso{
\link[MItools]{readInterProGFF3}
}
\author{
Vitalii Kleshchevnikov
}
