% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadMIontology.R
\name{loadMIontology}
\alias{loadMIontology}
\title{load Molecular Interaction ontology}
\usage{
loadMIontology(directory = "./", date = NULL,
  propagate_relationships = "is_a", extract_tags = "minimal")
}
\arguments{
\item{directory}{where to keep local copy of the MI ontology}

\item{date}{character, such as "2017_08_25", use specific version of ontology (local copy must be present), by default (NULL) download the latest version}

\item{propagate_relationships}{parameter for reading ontology into R, details: \code{\link[ontologyIndex]{get_ontology}}}

\item{extract_tags}{parameter for reading ontology into R, details: \code{\link[ontologyIndex]{get_ontology}}}
}
\value{
\code{MITABdata}
}
\description{
download and read Molecular Interaction ontology into R using \code{ontologyIndex} package
}
\author{
Vitalii Kleshchevnikov
}
