% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadTaxIDProteome.R
\name{loadTaxIDAllLower}
\alias{loadTaxIDAllLower}
\title{load all lower taxonomy ID (all descendants)}
\usage{
loadTaxIDAllLower(taxid, dir, with_description = F)
}
\arguments{
\item{file}{directory where to save/look for the local copy}
}
\value{
\code{loadTaxIDAllLower} returns list containing the character vector of all lower taxonomy ID for a given taxonomy ID (saves it to a file) and the input taxonomy ID (taxid) and optionally, a data.table containing the description of all lower
}
\description{
\code{loadTaxIDAllLower} loads all load all lower taxonomy ID for a given taxonomy ID
}
\examples{
# Find all human sub-species
loadTaxIDAllLower(taxid = 9606, dir = "./")
loadTaxIDAllLower(taxid = 9606, dir = "./", with_description = T)
}
\author{
Vitalii Kleshchevnikov

Vitalii Kleshchevnikov
}
