% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SLIMFinderOcc2GRanges.R
\name{merge2GRangesmcols}
\alias{merge2GRangesmcols}
\title{Merge data.table to GRanges metadata}
\usage{
merge2GRangesmcols(range, datatable, by.x = c("query", "interacts_with"),
  by.y = c("IDs_interactor_viral", "IDs_interactor_human"))
}
\arguments{
\item{range}{a path to a tsv (txt) file containing QSLIMFinder (SLIMFinder) occurence output}

\item{datatable}{a path to a tsv (txt) file containing QSLIMFinder (SLIMFinder) main output}

\item{one_from_cloud}{pick only one motif per motif cloud by the lowest p-value (before multiple hypothesis testing correction)}
}
\value{
Genomic Ranges object containing QSLIMFinder (SLIMFinder) occurence output
}
\description{
Merge data.table to GRanges metadata
}
\seealso{
\code{\link{ELMdb2GRanges}}, \code{\link{SLIMFinderOcc2GRanges}}
}
\author{
Vitalii Kleshchevnikov
}
