% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MITABregionFeature.R
\name{MITABregionFeature}
\alias{MITABregionFeature}
\title{Extract binding region feature from cleaned MITAB2.7}
\usage{
MITABregionFeature(mitab)
}
\arguments{
\item{mitab}{data.table containing data in MITAB2.7 format, standard column names should be made data.table-compatible}
}
\value{
data.table to which the following columns were added: binding_region_A_type, binding_region_A, binding_region_B_type, binding_region_B
}
\description{
Extract binding region feature from cleaned MITAB2.7
}
\author{
Vitalii Kleshchevnikov
}
