% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarkMotifs.R
\name{motifVariantClass}
\alias{motifVariantClass}
\title{Generate motif variant class (the first 2 element in the motif names)}
\usage{
motifVariantClass(ids)
}
\arguments{
\item{ids}{character vector of ELM motif IDs}
}
\value{
character vector of ELM motif variant class (such as TRG_NLS or LIG_SH3)
}
\description{
Generate motif variant class (the first 2 element in the motif names)
}
\author{
Vitalii Kleshchevnikov
}
