% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parReplicate.R
\name{parReplicate}
\alias{parReplicate}
\title{parReplicate function from: https://stackoverflow.com/questions/19281010/simplest-way-to-do-parallel-replicate}
\usage{
parReplicate(cl, n, expr, simplify = TRUE, USE.NAMES = TRUE)
}
\description{
parReplicate function from: https://stackoverflow.com/questions/19281010/simplest-way-to-do-parallel-replicate
}
\details{
arguments as in \code{\link{parSapply}} and \code{\link{replicate}}
}
