% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutationPval.R
\name{permutationPval}
\alias{permutationPval}
\title{calculate permutation pval for association between X and Z if both are connected to Y}
\usage{
res = permutationPval(interactions2permute = nodeX ~ nodeY,
 associations2test = nodeX ~ nodeZ,
 node_attr = NULL, data, statistic,
 select_nodes = NULL, N = 1000,
 cores = NULL, seed = NULL,
 also_permuteYZ = F)

# print
res

# plot p-value distribution (hist)
plot(res)

# plot the number of Y without Z per X (hist),
# formula is used to subset the table before plotting
# to avoid plotting single number multiple times
plot(res, nodeX ~ YmissingZ_perX)
}
\arguments{
\item{interactions2permute}{formula specifying columns that contain X-Y interactions that will be permuted}

\item{associations2test}{formula specifying columns that contain X-Z correspondence: empirical p-values for the association of X with Z will be tested, NOTE: Y-Z interactions are assumed}

\item{node_attr}{formula or list of formulas specifying columns that contain attributes of \code{X}, \code{Y} or \code{Z} or their combination (\code{X} ~ \code{degree}, \code{X} + \code{Z} ~ \code{pvalue})}

\item{data}{data.table containing interaction data and attributes}

\item{statistic}{formula that specifies how to calculate statistic using attibutes from \code{node_attr} by node \code{X} in \code{interactions2permute} and \code{Z} in \code{associations2test}, details: \code{\link[MItools]{permutationPvalHelper}}. In data.table synthax: \code{DT[, (observed/permuted)statistic := eval(right-hand-side of formula), by = .(eval(column names in the left-hand-side of formula))]}}

\item{select_nodes}{formula or list of formulas specifying which nodes of specific node type to select before permutation based on condition (\code{X} ~ \code{degree} > \code{10})}

\item{N}{number of times to run permutation of PPI network}

\item{cores}{specify how many cores to use for parallel processing, default (NULL) is to detect all cores on the machine and use all minus one. When using LSF cluster you must specify the number of cores to use because \code{\link[parallel]{detectCores}} doen't know how much cores you have requested from LSF (with bsub -n) and detects all cores on the actual physical node.}

\item{seed}{seed for RNG for reproducible sampling}

\item{also_permuteYZ}{logical, permute Y-Z interactions in addition to X-Y (specified in interactions2permute) ?}

\item{clustermq}{if TRUE uses clustermq job scheduling (\code{\link[clustermq]{Q}}) instead of local parallelisation (\code{\link[MItools]{parReplicate}})  = F,  = 4000,  = 100, clustermq_template = list(), split_comp_inner_N}

\item{clustermq_mem}{memory in MB to allocate for each job (ignored unless clustermq == TRUE)}

\item{clustermq_jobs}{maximal number of computing cluster jobs to use (ignored unless clustermq == TRUE)}

\item{clustermq_template}{Add specific arguments to computing cluster job submission call. Not needed in most cases. Details: \code{\link[clustermq]{Q}} (ignored unless clustermq == TRUE)}

\item{split_comp_inner_N}{parallel evaluation of permutations is split into the outer and inner replicate calls helps to save memory by decreasing the total size of the result. This argument let's you manually specify the number of inner replicate calls. This has to be optimised for data size when using clustermq (ignored unless clustermq == TRUE)}

\item{clustermq_fail_on_error}{If TRUE clustermq will fail if one of the jobs returns an error. Details: \code{\link[clustermq]{Q}} (ignored unless clustermq == TRUE)}

\item{formula}{argument for \code{permutationPvalPlot}, formula specifying attribute of which nodes to plot like this: nodeX + nodeZ ~ p.value. The default is to plot p.value histogram for nodeX and nodeZ as specified in the \code{x} object}

\item{x}{argument for \code{permutationPvalPlot}, output of \code{permutationPval}, class "XYZinteration_XZEmpiricalPval"}

\item{...}{argument for \code{permutationPvalPlot}, base R plotting parameters}
}
\value{
object of S3 class "XYZinteration_XZEmpiricalPval" (list), containing \code{permutationPval} function call, standardised node names, and data.table containing the original data but appended with empirical p-value (p.value), observed_statistic, YmissingZ_perX, and higher_counts, not_missing used to calculate p-value
}
\description{
calculate permutation pval for association between X and Z if both are connected to Y
}
\author{
Vitalii Kleshchevnikov
}
