% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula2colnames.R
\docType{data}
\name{permutationPvalHelper}
\alias{permutationPvalHelper}
\alias{formula2colnames}
\alias{node_attr2colnames}
\title{functions that are called inside permutationPval}
\format{An object of class \code{NULL} of length 0.}
\usage{
permutationPvalHelper

cols = formula2colnames(node_attr, cols, nodes)
cols = node_attr2colnames(node_attr, cols, nodes)
nodes = formula2nodes(interactions2permute, associations2test)
data_list = filterByFormula(data_list, select_nodes, cols, nodes)
data_list = calcObservedStatistic(data_list, by_cols, exprs, nodes, nodes_call, includeAssociations)
data_list = calcPermutedStatistic(data_list, by_cols, exprs, nodes, nodes_call, includeAssociations, also_permuteYZ)
res = observedVSpermuted(data_list, nodes_call, nodes)
temp2 = aggregatePermutations(temp, nodes, nodes_call)
}
\arguments{
\item{node_attr}{formula (formula2colnames and node_attr2colnames) or list of formulas (node_attr2colnames)}

\item{cols}{list of column names}

\item{nodes}{list of node names (column names for columns that contain node names)}

\item{nodes_call}{list of node names, each class call (to be evaluated within data.table expressions DT[,eval(nodes_call$nodeX)])}

\item{interactions2permute}{decription: \code{\link[MItools]{permutationPval}}}

\item{associations2test}{decription: \code{\link[MItools]{permutationPval}}}

\item{data_list}{list (or enviroment?) that contains data.tables necessary for these functions to work}

\item{select_nodes}{decription: \code{\link[MItools]{permutationPval}}}

\item{by_cols}{modified left-hand side of \code{statistic} (\code{\link[MItools]{permutationPval}}), class call}

\item{exprs}{right-hand side of \code{statistic} (\code{\link[MItools]{permutationPval}}), class call, \code{exprs} is evaluated by \code{by_cols}; in data.table synthax: DT[, statistic := eval(exprs), by = .(eval(by_cols))]}

\item{includeAssociations}{logical, if calculating statistic requires columns that contain attribute of X-Z pair associations table will be merged}

\item{also_permuteYZ}{logical, permute Y-Z interactions as well as X-Y ?}
}
\value{
updated list of column names
}
\description{
functions that are called inside permutationPval

extract and match attribute column names to data.tables that contain nodes which those attributes describe
}
\details{
extract nodes

extract attributes

attach attributes to a table if all nodes are present in that table

and in case of associations table both varibles in a formula should be nodeX and nodeZ
}
\author{
Vitalii Kleshchevnikov
}
\keyword{datasets}
