% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PermutResult2D.R
\name{PermutResult2D}
\alias{PermutResult2D}
\title{2D bin plot of the features of top N domain-protein pairs as identified by \code{\link[MItools]{permutationPval}}}
\usage{
PermutResult2D(res, N, id.cols = c("IDs_interactor_viral",
  "IDs_domain_human"), value.cols = c("domain_count",
  "IDs_interactor_viral_degree", "domain_count_per_IDs_interactor_viral",
  "p.value"), rank_by = "p.value", filter = NULL)
}
\arguments{
\item{res}{\code{\link[MItools]{permutationPval}} output of (class XYZinteration_XZEmpiricalPval)}

\item{N}{the number of top pairs to look at}

\item{id.cols}{character, columns that define domain-protein pairs (domain id and protein id)}

\item{value.cols}{which values to plot}

\item{top_by}{column name by which to select top N pair}
}
\value{
ggplot
}
\description{
2D bin plot of the features of top N domain-protein pairs as identified by \code{\link[MItools]{permutationPval}}
}
\author{
Vitalii Kleshchevnikov
}
