% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEnrichment.R
\name{plotEnrichment}
\alias{plotEnrichment}
\title{Plot the enrichment of domains that are likely to be mediating interactions in domain annotated by ELM as mediating linear motif mediated interactions}
\usage{
plotEnrichment(..., list = list(), random_domains = NULL,
  domains_known_mapped, type = "count", plot_type = "l", plot_name = "")
}
\arguments{
\item{...}{objects (matrix) the output of \code{runningTestEnrichment}}

\item{list}{list of object (alternative way to pass the output of \code{runningTestEnrichment})}

\item{random_domains}{the result of comparing random domain-protein pairs (by \code{random_domains} function), can be null}

\item{domains_known_mapped}{domains from ELM mapped to InterProID}

\item{type}{what to plot: "pval"- Fisher test p-val (in out set vs in ELM), "odds_ratio" - Fisher test odds ratio, "count" of domains in our set that are in ELM}

\item{plot_type}{r-base plot type default is "l"}
}
\value{
plot
}
\description{
Plot the enrichment of domains that are likely to be mediating interactions in domain annotated by ELM as mediating linear motif mediated interactions
}
\author{
Vitalii Kleshchevnikov
}
