% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printMotifDomainTable.R
\name{printMotifDomainTable}
\alias{printMotifDomainTable}
\title{\code{printMotifDomainTable} combines Domain enrichment and Motif search results}
\usage{
printMotifDomainTable(input, doman_viral_pairs = F, motifs = F,
  destfile, fdr_pval_thresh = 0.05, only_with_motifs = F,
  fdr_motifs = 1, occurence_QSLIMFinder = NA, comparimotif_wdb = NA,
  patterns_QSLIMFinder = NA, print_table = T, one_from_cloud = T,
  entry.list = NULL)
}
\arguments{
\item{input}{object of class XYZinteration_XZEmpiricalPval and should be the output of permutationPval()}

\item{doman_viral_pairs}{if TRUE - remove human proteins to have a row per viral protein - human domain pair, if FALSE a row is per viral protein - human domain - human protein.}

\item{motifs}{add motif search results?}

\item{destfile}{file where to save the resulting table}

\item{fdr_pval_thresh}{FDR threshold p-value for Domain enrichment analysis}

\item{only_with_motifs}{show enriched domains only if motif in viral proteins are also present (if FALSE - show all results)}

\item{fdr_motifs}{FDR threshold p-value for Motif search analysis}

\item{occurence_QSLIMFinder}{data.table containing the motif occurence output of the QSLIMFinder}

\item{comparimotif_wdb}{data.table containing the output of the comparimotif tool (comparing motifs from \code{occurence_QSLIMFinder} to ELM DB)}

\item{print_table}{if TRUE calls DT:datatable to print the interactive results table}

\item{one_from_cloud}{pick only one motif per motif cloud by the lowest p-value (before multiple hypothesis testing correction)}
}
\description{
\code{printMotifDomainTable} combines Domain enrichment and Motif search results
}
\author{
Vitalii Kleshchevnikov
}
