% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QSLIMFinderCommand.R
\name{QSLIMFinderCommand}
\alias{QSLIMFinderCommand}
\alias{mQSLIMFinderCommand}
\title{Generate a QSLIMFinder command}
\usage{
QSLIMFinderCommand(file_list, i = 1,
  slimpath = "../software/slimsuite/tools/qslimfinder.py",
  blast = "../software/ncbi_blast_2.6.0/bin/",
  iupred = "../software/iupred/iupred",
  options = "dismask=T consmask=F cloudfix=F probcut=0.1",
  LSF_cluster_par = "bsub -n 1 -q research-rh7 -M 100 -R \\"rusage[mem=100]\\"",
  LSF_project_path = "/hps/nobackup/research/petsalaki/users/vitalii/vitalii/viral_project/")

mQSLIMFinderCommand(file_list,
  slimpath = "../software/slimsuite/tools/qslimfinder.py",
  blast = "../software/ncbi_blast_2.6.0/bin/",
  iupred = "../software/iupred/iupred",
  options = "dismask=T consmask=F cloudfix=F probcut=0.5",
  LSF_cluster_mode = F,
  LSF_project_path = "/hps/nobackup/research/petsalaki/users/vitalii/vitalii/viral_project/",
  LSF_cluster_par = "bsub -n 1 -q research-rh7 -M 200 -R \\"rusage[mem=200]\\"",
  log_dir = "./SLIMFinder/log_dir/", write_log = T, recursive = F)
}
\arguments{
\item{file_list}{data.table containing path to files and directories for QSLIMFinder: fastafile, queryfile, outputdir, outputfile}

\item{i}{integer, which set of files and directories to choose from \code{file_list}}

\item{slimpath}{relative path (from the project folder) to qslimfinder.py}

\item{blast}{relative path (from the project folder) to /bin/ folder in BLAST package}

\item{iupred}{relative path (from the project folder) to iupred (compiled executable)}

\item{options}{any options from QSLIMFinder}

\item{LSF_cluster_par}{a string that will launch LSF job}

\item{LSF_project_path}{absolute path on LSF cluster}

\item{LSF_cluster_mode}{logical, if FALSE \code{LSF_cluster_par} and \code{LSF_project_path} are ignored}

\item{log_dir}{character, directory where to write log files (stout and sterr)}

\item{write_log}{whether to save stout and sterr}
}
\value{
character vector (1L), bash command that will lauch QSLIMFinder locally or as a job on LSF cluster

list containing: 1. command to set up enviromental variable IUPred_PATH; 2. character vector of bash commands that will lauch QSLIMFinder as a job on LSF cluster; 3, 4, 5 - directories where LSF should write stout and sterr
}
\description{
Generate a QSLIMFinder command
}
\examples{
all_commands = mQSLIMFinderCommand(file_list = forSLIMFinder_file_list,
    slimpath = "../software/cluster/slimsuite/tools/qslimfinder.py",
    blast = "../software/cluster/ncbi_blast_2.6.0/bin/",
    iupred = "../software/cluster/iupred/iupred",
    options = "dismask=T consmask=F cloudfix=T probcut=0.3 iuchdir=T",
    LSF_cluster_mode = T,
    LSF_project_path = "/hps/nobackup/research/petsalaki/users/vitalii/vitalii/viral_project/",
    LSF_cluster_par = "bsub -n 1 -q research-rh7 -M 100 -R \\"rusage[mem=100]\\"")
}
\seealso{
\code{\link{listInteractionSubsetFASTA}}, \code{\link{runQSLIMFinder}}
}
\author{
Vitalii Kleshchevnikov
}
