% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryPSICQUIC.R
\name{queryPSICQUIC}
\alias{queryPSICQUIC}
\title{Use PSICQUIC to query for molecular interactions (using MIQL, specifying database and output format)}
\usage{
queryPSICQUIC(query, format, database, file)
}
\arguments{
\item{query}{search query in MIQL query language, as you would type in PSICQUIC View client: \url{http://www.ebi.ac.uk/Tools/webservices/psicquic/view/home.xhtml}}

\item{format}{output format (most widely used tabular format is tab25, use tab27 for more data columns)}

\item{database}{PSICQUIC service (full list: \url{http://www.ebi.ac.uk/Tools/webservices/psicquic/registry/registry?action=STATUS}), use "imex" shorthand to choose all IMEx databases}

\item{file}{file name and directory to save the result ("/dir/filename")}
}
\value{
saves query result to a file, returns the query settings, how many interactions per database retrieved and which databases are inactive
}
\description{
query_PSICQUIC function queries a specified molecular interaction database using MIQL query language, retrieves the data in a specified format and saves it to a file
}
\details{
See \url{https://psicquic.github.io/MiqlReference.html} or \url{https://psicquic.github.io/MiqlReference27.html} for the description of MIQL query language.

Unlike rawQuery function from PSICQUIC package this function allows to use all IMEx databases by passing "imex" to database argument and also separates the choice of output format from the query. Output format options: \url{https://psicquic.github.io/formats.html}

This function also splits large query result into chunks (2500 interactions) to limit the load on the PSICQUIC servers.

List of data provider names accepted by the function: "imex", "APID Interactomes", "BioGrid", "bhf-ucl", "ChEMBL", "HPIDb", "InnateDB", "InnateDB-All", "IntAct", "mentha", "MPIDB", "MatrixDB", "MINT", "Reactome", "Reactome-FIs", "EBI-GOA-miRNA", "I2D", "I2D-IMEx", "InnateDB-IMEx", "MolCon", "UniProt", "MBInfo", "VirHostNet", "BAR", "EBI-GOA-nonIntAct", "ZINC"
}
\examples{
queryPSICQUIC(query = "id:P74565 AND detmethod:\\"MI:0018\\"",
               format = "tab27",
               database = "imex",
               file = "P74565_2H_interactions_imex_tab27.tsv")

queryPSICQUIC(query = "id:P74565 AND detmethod:\\"MI:0018\\"",
               format = "tab25",
               database = "mentha",
               file = "P74565_2H_interactions_mentha_tab25.tsv")

queryPSICQUIC(query = "id:P74565",
               format = "tab25",
               database = "mentha",
               file = "P74565_2H_interactions_mentha_tab25.tsv")

queryPSICQUIC(query = "id:156",
               format = "tab25",
               database = "BioGrid",
               file = "entrezgene156_interactions_BioGrid_tab25.tsv")

}
\author{
Vitalii Kleshchevnikov

Vitalii Kleshchevnikov
}
