% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomGRanges.R
\name{randomGRanges}
\alias{randomGRanges}
\title{Generate random GRanges of equal length in the same sequence}
\usage{
randomGRanges(GRanges, N = 1, replace = T, within1sequence = T)
}
\arguments{
\item{GRanges}{GRanges object}

\item{N}{how many randomised GRanges to create}

\item{replace}{passed to \code{\link[base]{sample.int}}}

\item{within1sequence}{logical, if TRUE ranges of equal lenght are sampled from the same sequence. If false, ranges of equal length are sampled from all sequences (while keeping the number of ranges in each sequence).}
}
\value{
Genomic Ranges list object containing \code{N} randomised GRanges objects
}
\description{
Generate random GRanges of equal length in the same sequence
}
\seealso{
\code{\link{ELMdb2GRanges}}, \code{\link{GRangesINinteractionSubsetFASTA}}
}
\author{
Vitalii Kleshchevnikov
}
