% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterProEntryTypes.R
\name{readInterProGFF3}
\alias{readInterProGFF3}
\alias{addInterProEntryTypes}
\title{Download FASTA for a list of canonical (P04637) and isoform (P04637-2) UniprotAC}
\usage{
addInterProEntryTypes(InterProScan_result, entry.list_path)
}
\arguments{
\item{InterProScan_result}{GRanges, the output of \code{readInterProGFF3} function}

\item{entry.list_path}{character, file name and directory where to store InterPro sca}

\item{filename.gz}{character, file name and directory to read InterProScan output in the gff format ("./dir/filename.gz")}

\item{selected_ENTRY_TYPE}{character, select InterPro signature matches from \code{InterProScan_result} using these InterPro Entry Types}
}
\value{
\code{readInterProGFF3}, \code{addInterProEntryTypes} or \code{SubsetByInterProEntryType}: \link[GenomicRanges]{GRanges-class} object containing InterProScan output appended by InterPro Entry Types information, with names metadata and sequence length information imported correctly

\code{getInterProEntryTypes}: data.table containing InterPro Entry List \link{ftp://ftp.ebi.ac.uk/pub/databases/interpro/entry.list}

\code{getInterPro2memberDB}: data.table containing  InterProID to member database ID mapping
}
\description{
Download FASTA for a list of canonical (P04637) and isoform (P04637-2) UniprotAC
}
\details{
For further details on the InterProScan output format please visit the link \link{https://github.com/ebi-pf-team/interproscan/wiki/OutputFormats}
}
\examples{
addInterProEntryTypes(filename.gz = "./processed_data_files/all_human_viral_protein_domains.gff3.gz")
}
\author{
Vitalii Kleshchevnikov
}
