% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readQSLIMFinderMain.R
\name{readQSLIMFinderMain}
\alias{readQSLIMFinderMain}
\alias{readQSLIMFinderOccurence}
\alias{writePatternList}
\title{Read and write QSLIMFinder results}
\usage{
readQSLIMFinderMain(outputfile = "forSLIMFinder_file_list$outputfile",
  check_exist = F)


  readQSLIMFinderOccurence(outputdir = "forSLIMFinder_file_list$outputdir",
  check_exist = F)

writePatternList(QSLIMFinder_main_result, filename = "./motifs.txt")
}
\arguments{
\item{outputfile}{character vector of paths to QSLIMFinder main output files, one way to get these is \code{\link{writeInteractionSubsetFASTA_list}}()$outputfile}

\item{check_exist}{if TRUE, check that each file exists / is not empty, if false - read all existing and non-empty files}

\item{outputdir}{character vector of paths to QSLIMFinder output directories, one way to get these is \code{\link{writeInteractionSubsetFASTA_list}}()$outputdir}

\item{QSLIMFinder_main_result}{data.table, the output of readQSLIMFinderMain()}

\item{filename}{character, paths to file where to write motif list}
}
\value{
readQSLIMFinderMain: data.table combined from all QSLIMFinder main output files. Contains summary statistic on each motif found in each dataset

readQSLIMFinderOccurence: data.table combined from all QSLIMFinder occurence output files. Contains positions of motifs in each protein.

writePatternList: data.table containing Name and Pattern columns is saved as a tsv to \code{filename}, a format required for by comparimotif
}
\description{
Read and write QSLIMFinder results

Read and write QSLIMFinder results

Read and write QSLIMFinder results
}
\seealso{
\code{\link{writeInteractionSubsetFASTA_list}}, \code{\link{runQSLIMFinder}}
}
\author{
Vitalii Kleshchevnikov

Vitalii Kleshchevnikov

Vitalii Kleshchevnikov
}
