% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mInteractionSubsetFASTA.R
\name{recodeANDinteractionSubsetFASTA}
\alias{recodeANDinteractionSubsetFASTA}
\title{Recode names and retrieve FASTA sequences of interactors of a specific protein}
\usage{
recodeANDinteractionSubsetFASTA(subset1, fasta)
}
\arguments{
\item{subset1}{interaction_subset_from_2_sets object produced by \code{\link{subset2setsBy1ID}}}

\item{fasta}{AAStringSet, names are UniProtKB accessions}
}
\value{
object of class interactionSubsetFASTA containing: FASTA sequences for interacting proteins, original interaction_subset_from_2_sets
}
\description{
Recode names and retrieve FASTA sequences of interactors of a specific protein
}
\examples{
subset1_fasta = recodeANDinteractionSubsetFASTA(subset1, fasta)
}
\seealso{
\code{\link{subset2setsBy1ID}}, \code{\link{recodeFASTA}}, \code{\link{listInteractionSubsetFASTA}}
}
\author{
Vitalii Kleshchevnikov
}
