% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recodeFASTA.R
\name{recodeFASTA}
\alias{recodeFASTA}
\title{Recode names of the AAStringSet object}
\usage{
recodeFASTA(fasta, mapping = NULL, i = NULL)
}
\arguments{
\item{fasta}{AAStringSet object for which to generate new names or the interactionSubsetFASTA which to rename back to the original names}

\item{mapping}{data.table, mapping table between new and original names}

\item{i}{which element of the interactionSubsetFASTA to rename back to the original names}
}
\value{
list of the original AAStringSet object with new names and mapping table between new and original names

interactionSubsetFASTA object with original names as provided by mapping table
}
\description{
Recode names of the AAStringSet object
}
\examples{
# from original to new
recoded.fasta = recodeFASTA(fasta)
# from new to original
subset_fasta = recodeFASTA(subset_fasta, recoded.fasta$names_mapping, i = 1)
}
\seealso{
\code{\link{downloadFasta_postproc}}
}
\author{
Vitalii Kleshchevnikov
}
