% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarkMotifs.R
\name{reduceWithMcols}
\alias{reduceWithMcols}
\title{Reduce GRanges keeping metadata (collapsing withing each column separated by pipes)}
\usage{
reduceWithMcols(range)
}
\arguments{
\item{range}{any Genomic Ranges object with metadata (mcols())}
}
\value{
reduced Genomic Ranges object
}
\description{
Reduce GRanges keeping metadata (collapsing withing each column separated by pipes)
}
\author{
Vitalii Kleshchevnikov
}
