% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadTaxIDProteome.R
\name{removeInteractionNoFASTA}
\alias{removeInteractionNoFASTA}
\title{Remove interactions if one participant doesn't have FASTA sequence}
\usage{
removeInteractionNoFASTA(MITABdata, fasta)
}
\arguments{
\item{MITABdata}{object of any clean_MItab class (the class that contains "clean_MItab" in it's name and is initially produced by \code{\link[PItools]{cleanMITAB}})}

\item{fasta}{AAStringSet, names are UniProtKB accessions}
}
\description{
filter molecular interaction data by the list of FASTA sequences in AAStringSet
}
\author{
Vitalii Kleshchevnikov
}
