% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadTaxIDProteome.R
\name{removeInteractionObsoleteID}
\alias{removeInteractionObsoleteID}
\title{Remove interactions if one participant has obsolete Uniprot accession (ID)}
\usage{
removeInteractionObsoleteID(MITABdata, dir = "./")
}
\arguments{
\item{MITABdata}{object of any clean_MItab class (the class that contains "clean_MItab" in it's name and is initially produced by \code{\link[MItools]{cleanMITAB}})}

\item{dir}{directory where to save/look for the local copy}
}
\description{
filter molecular interaction data by Pubmed ID of publications
}
\author{
Vitalii Kleshchevnikov
}
