% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeSelfLoops.R
\name{removeSelfLoops}
\alias{removeSelfLoops}
\title{Remove self loops (self-interactions)}
\usage{
removeSelfLoops(MITABdata)
}
\arguments{
\item{MITABdata}{molecular interaction data, object of class "clean_MItab" (or any other object which names starts from "clean_MItab")}
}
\value{
molecular interaction data with the self interactions removed
}
\description{
subset molecular interaction data with a list of interactors
}
\seealso{
\code{\link[MItools]{subsetMITABbyMethod}}), \code{\link[MItools]{subsetMITABbyPMIDs}})
}
\author{
Vitalii Kleshchevnikov
}
