% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanMITAB25.R
\name{reorderMITAB25}
\alias{reorderMITAB25}
\title{\code{reorderMITAB25} reorders interacting molecules in a pair (and all the corresponding columns) according to order provided in IDs_A_order and IDs_B_order columns (latter are deleted)}
\usage{
reorderMITAB25(mitab)
}
\arguments{
\item{mitab}{data.table containing molecular interaction data in MITAB 2.5 format}
}
\description{
function called internally by \code{\link{cleanMITAB25}} if the format is MITAB2.5
}
\author{
Vitalii Kleshchevnikov
}
