% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanMITAB27.R
\name{reorderMITAB27}
\alias{reorderMITAB27}
\title{\code{reorderMITAB27} extracts and cleans a set of columns of MITAB2.7}
\usage{
reorderMITAB27(mitab)
}
\arguments{
\item{mitab}{data.table containing molecular interaction data in MITAB 2.7 format}
}
\description{
function called internally by \code{\link{cleanMITAB}} if the format is MITAB2.7
}
\author{
Vitalii Kleshchevnikov
}
