% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCompariMotif3.R
\name{runCompariMotif3}
\alias{runCompariMotif3}
\title{Run CompariMotif3}
\usage{
runCompariMotif3(input_file = "./SLIMFinder/result/motifs.txt",
  slimpath = "../software/slimsuite/tools/", dbpath = "./data_files/",
  dburl = "http://elm.eu.org/elms/elms_index.tsv",
  parameters = "unmatched=T motific=T",
  out_file = "./SLIMFinder/result/comparimotif.tdt",
  LSF_project_path = "/hps/nobackup/research/petsalaki/users/vitalii/vitalii/viral_project/",
  run = F, with = c("self", "db")[2])
}
\arguments{
\item{input_file}{a path to a tsv file containing Name and Pattern columns}

\item{slimpath}{path to /slimsuite/tools/}

\item{dbpath}{path to directory where to save and keep ELM database (\link{http://elm.eu.org/}) or other database of linear motifs in a format required by comparimotif_V3: \link{http://rest.slimsuite.unsw.edu.au/docs&page=module:comparimotif_V3}}

\item{dburl}{url where to download database}

\item{parameters}{any commandline parameters for comparimotif_V3 as a single character string}

\item{out_file}{path to a file where comparimotif_V3 should save the result}

\item{LSF_project_path}{path to the project directory on the LSF cluster file system. If on LSF is to be used all other paths should be relative to this directory, alternatively, set this argument to empty character. Final path is \code{LSF_project_path} concatenated with \code{input_file}, \code{slimpath}, \code{out_file} or \code{dbpath}. When using this on a local machine set this argument to empty character.}

\item{run}{logical, run comparimotif_V3 (T) or just output the command (F), can be useful for check correcteness}

\item{with}{compare motifs in \code{input_file} to the database in \code{dbpath} ("db") or to \code{input_file} ("self")}
}
\value{
system command to be run or that was run
}
\description{
Run CompariMotif3
}
\seealso{
\code{\link{writeInteractionSubsetFASTA_list}}, \code{\link{runQSLIMFinder}}
}
\author{
Vitalii Kleshchevnikov
}
