% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mInteractionSubsetFASTA.R
\name{singleInteractFromSet2}
\alias{singleInteractFromSet2}
\title{Filter interaction subset from 2 sets: keep only i-th element from the set2}
\usage{
singleInteractFromSet2(subset1, set1_only, i)
}
\arguments{
\item{subset1}{interaction_subset_from_2_sets object produced by \code{\link{subset2setsBy1ID}}}

\item{set1_only}{logical, only relevant if \code{single_interact_from_set2 = TRUE}, sequence set1 should contain only proteins that interact with an element of seed_id_vect in interaction_set1. If FALSE, proteins that interact with an element of seed_id_vect in interaction_set2 with the exception of the single one () are also included}

\item{i}{which protein in the set2 inside \code{subset1} to keep as a set2 element}
}
\value{
object of class interaction_subset_from_2_sets, the same as returned by \code{\link{subset2setsBy1ID}}, but set2 contains only one protein and (optionally) all other set2 proteins and corresponding interactions are removed
}
\description{
Filter interaction subset from 2 sets: keep only i-th element from the set2
}
\examples{
subset1_1 = singleInteractFromSet2(subset1, set1_only = set1_only, i = 1)
}
\seealso{
\code{\link{subset2setsBy1ID}}, \code{\link{recodeFASTA}}, \code{\link{listInteractionSubsetFASTA}}
}
\author{
Vitalii Kleshchevnikov
}
