% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SLIMFinderOcc2GRanges.R
\name{SLIMFinderOcc2GRanges}
\alias{SLIMFinderOcc2GRanges}
\title{Read QSLIMFinder (SLIMFinder) occurence output to Genomic Ranges object}
\usage{

  SLIMFinderOcc2GRanges(occurence_file = "../viral_project/SLIMFinder_Vidal/result/occurence.txt",
  main_file = "../viral_project/SLIMFinder_Vidal/result/main_result.txt",
  one_from_cloud = T)
}
\arguments{
\item{main_file}{a path to a tsv (txt) file containing QSLIMFinder (SLIMFinder) main output}

\item{one_from_cloud}{pick only one motif per motif cloud by the lowest p-value (before multiple hypothesis testing correction)}

\item{file}{a path to a tsv (txt) file containing QSLIMFinder (SLIMFinder) occurence output}
}
\value{
Genomic Ranges object containing QSLIMFinder (SLIMFinder) occurence output
}
\description{
Read QSLIMFinder (SLIMFinder) occurence output to Genomic Ranges object
}
\seealso{
\code{\link{ELMdb2GRanges}}, \code{\link{GRangesINinteractionSubsetFASTA}}
}
\author{
Vitalii Kleshchevnikov
}
