% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset2setsBy1ID.R
\name{subset2setsBy1ID}
\alias{subset2setsBy1ID}
\title{Subset 2 sets of protein interaction data using a single ID and combine the results}
\usage{
subset2setsBy1ID(interaction_set1, interaction_set2, seed_id)
}
\arguments{
\item{interaction_set1}{cleanMItab object containing molecular interactions}

\item{interaction_set2}{different cleanMItab object containing molecular interactions}

\item{seed_id}{interactor ID for which to retrieve interactions}
}
\value{
object of class interaction_subset_from_2_sets (list): name = seed_id,

combined_MITAB = combined interaction data for seed_id from interaction_set1 and interaction_set2

ids_all = IDs of all interactors of seed_id

ids_set1 = IDs of interactors of seed_id in interaction_set1

ids_set2 = IDs of interactors of seed_id in interaction_set1

length_set1, length_set2 - number of interactions in each set

description - short decription of the data
}
\description{
Subset 2 sets of protein interaction data using a single ID and combine the results
}
\examples{
subset = subset2setsBy1ID(interaction_set1 = all_human_interaction,
                        interaction_set2 = all_viral_interaction,
                        seed_id = proteins_w_signif_domains[2])
# print the result
subset
}
\seealso{
\code{\link{interactionSubsetMapID}}
}
\author{
Vitalii Kleshchevnikov
}
