% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetMITABbyID.R
\name{subsetMITABbyID}
\alias{subsetMITABbyID}
\title{Retrieve interactions for a list of interactors (given data table, not using webservice)}
\usage{
subsetMITABbyID(MITABdata, ID_seed, within_seed = F,
  only_seed2nonseed = F)
}
\arguments{
\item{MITABdata}{molecular interaction data, object of class "clean_MItab25", "clean_MItab27", "clean_MItab25_fullInteractome", "clean_MItab27_fullInteractome"}

\item{ID_seed}{filter \code{MITABdata} using this list of interactors}

\item{within_seed}{logical, should \code{subsetMITABbyID} return only interactions between molecules in \code{ID_seed} (TRUE) or any interactions molecules in \code{ID_seed} have in \code{MITABdata}}
}
\value{
molecular interaction data, object of class `class of MITABdata`_subset (for example, clean_MItab27_subset, basically list), a subset of \code{MITABdata} that contains interactions for \code{ID_seed} molecules, and additionally includes the ID_seed and within_seed
}
\description{
subset molecular interaction data with a list of interactors
}
\seealso{
\code{\link[PItools]{subsetMITABbyMethod}}), \code{\link[PItools]{subsetMITABbyPMIDs}})
}
\author{
Vitalii Kleshchevnikov
}
