% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetMITABbyMethod.R
\name{subsetMITABbyMethod}
\alias{subsetMITABbyMethod}
\title{Filter molecular interactions by interaction detection methods and / or participant identification method}
\usage{
subsetMITABbyMethod(MITABdata, Interaction_detection_methods = NULL,
  Identification_method = NULL,
  Identification_method_participant_A = NULL,
  Identification_method_participant_B = NULL,
  ontology_directory = "./", ontology_date = NULL,
  inverse_filter = F)
}
\arguments{
\item{MITABdata}{object of any clean_MItab class (the class that contains "clean_MItab" in it's name and is initially produced by  \code{\link[MItools]{MITABdata}})}

\item{Interaction_detection_methods}{character or character vector, Molecular Interaction ontology code for Interaction detection method, such as "MI:0018", default option results in no filtering by Interaction detection method. Details \link{http://purl.obolibrary.org/obo/MI_0001}}

\item{Identification_method}{character or character vector, Molecular Interaction ontology code for Identification method, such as "MI:0433". Applies the same Identification method filter to both participant_A and participant_B. Default option results in no filtering by Identification method. Details \link{http://purl.obolibrary.org/obo/MI_0002}}

\item{Identification_method_participant_A}{character or character vector, Molecular Interaction ontology code for Identification method for participant A, such as "MI:0433", default option results in no filtering by Interaction detection method for this participant. Parameter is ignored if \code{Identification_method} specified. Details \link{http://purl.obolibrary.org/obo/MI_0002}}

\item{Identification_method_participant_B}{character or character vector, Molecular Interaction ontology code for Identification method for participant B, such as "MI:0433", default option results in no filtering by Interaction detection method for this participant. Parameter is ignored if \code{Identification_method} specified. Details \link{http://purl.obolibrary.org/obo/MI_0002}}

\item{ontology_directory}{where local copy of the MI ontology is stored (if file not found it will be downloaded)}

\item{ontology_date}{character, such as "2017_08_25", use specific version of ontology (local copy must be present), by default (NULL) download the latest version. The function doesn't require internet access if this argument is provided.}

\item{inverse_filter}{logical, inverse filtering criteria}
}
\value{
object of the same class as \code{MITABdata}, a subset of \code{MITABdata} that contains interactions for \code{Interaction_detection_methods} and / or participant \code{Identification_method}
}
\description{
filter molecular interaction data by \code{Interaction_detection_methods} and / or participant \code{Identification_method}
}
\examples{
# Download all human interaction data
full = fullInteractome(taxid = "9606", database = "IntActFTP",
         clean = TRUE, protein_only = TRUE)

# subset two-hybrid interactions
two_hybrid = subsetMITABbyMethod(MITABdata = full,
               Interaction_detection_methods = "MI:0018")

# subset all interactions but two-hybrid
NOT_two_hybrid = subsetMITABbyMethod(MITABdata = full,
                   Interaction_detection_methods = "MI:0018", inverse_filter = T)

# subset affinity purification - mass spectrometry interactions
AP_MS = subsetMITABbyMethod(MITABdata = full,
               Interaction_detection_methods = "MI:0004",  Identification_method = "MI:0433")
}
\seealso{
\code{\link[MItools]{subsetMITABbyPMIDs}}), \code{\link[MItools]{subsetMITABbyID}})
}
\author{
Vitalii Kleshchevnikov
}
