% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeInteractionSubsetFASTA_list.R
\name{writeInteractionSubsetFASTA_list}
\alias{writeInteractionSubsetFASTA_list}
\title{Write InteractionSubsetFASTA_list into directory system (creating directories as necessary) and return file locations for QSLIMFinder}
\usage{
writeInteractionSubsetFASTA_list(interactionFASTA_list, dir = "./SLIMFinder/")
}
\arguments{
\item{interactionFASTA_list}{object of class InteractionSubsetFASTA_list created by \code{\link{listInteractionSubsetFASTA}}, and (potentially) filtered using \code{\link{filterInteractionSubsetFASTA_list}} or \code{\link{domainProteinPairMatch}}}

\item{dir}{directory where to write interactionFASTA_list}
}
\value{
data.table containing path to files and directories for QSLIMFinder: fastafile, queryfile, outputdir, outputfile
}
\description{
Write InteractionSubsetFASTA_list into directory system (creating directories as necessary) and return file locations for QSLIMFinder
}
\examples{
forSLIMFinder_file_list = writeInteractionSubsetFASTA_list(interactionFASTA_list = forSLIMFinder_Ready,
                           dir = "./SLIMFinder/")
}
\seealso{
\code{\link{listInteractionSubsetFASTA}}
}
\author{
Vitalii Kleshchevnikov
}
