% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printMotifDomainTable.R
\name{XYZ.p.adjust}
\alias{XYZ.p.adjust}
\title{\code{XYZ.p.adjust} adjusts p.value in the permutation result}
\usage{
XYZ.p.adjust(res, adj_by = "p.value", ...)
}
\arguments{
\item{res}{object of class XYZinteration_XZEmpiricalPval and should be the output of permutationPval()}

\item{adj_by}{column name that contains p-values}

\item{...}{passed to p.adjust()}
}
\description{
\code{XYZ.p.adjust} adjusts p.value in the permutation result
}
\author{
Vitalii Kleshchevnikov
}
