% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarkMotifsROC.R
\name{benchmarkMotifsROC}
\alias{benchmarkMotifsROC}
\alias{mBenchmarkMotifsROC}
\title{Plot ROC or precision-recall curves for motifs search benchmark}
\usage{
benchmarkMotifsROC(res, data_type = c("both", "query", "all")[1],
  col_query = c("grey"), col_all = c("black"), ...,
  text_args = "col = \\"black\\"", ROCR_data = F,
  legend_args = "x = 0.8 | y = 0.25", x_query = 0.25, x_all = 0.75,
  measure1 = "tpr", measure2 = "fpr", single_metric = c("auc",
  "prbe")[1], single_metric_name = c("Median AUC",
  "Precision-recall break-even point")[1],
  summarise_single_metric = c("median", "")[1],
  both_metrics_vs_cutoff = F)

mBenchmarkMotifsROC(res_list, data_type = "query", ...)
}
\arguments{
\item{res}{object class \code{benchmarkMotifsResult}, the output of \code{\link{benchmarkMotifs}}}

\item{...}{args passed to \code{\link[ROCR]{plot}}}

\item{text_args}{character 1L, arguments to \code{\link[base]{text}} and \code{\link[base]{legend}} that allow to manipulate the median AUC and legend on the plot}

\item{ROCR_data}{logical, return ROC performance data? If FALSE the function only plots ROC curve}

\item{legend_args}{character 1L, arguments to \code{\link[base]{text}} and \code{\link[base]{legend}} that allow to manipulate the median AUC and legend on the plot}

\item{measure1}{character, for example "tpr" or "prec", details: ?ROCR::performance}

\item{measure2}{character, for example "fpr" or "rec", details: ?ROCR::performance}

\item{res_list}{list of objects of class \code{benchmarkMotifsResult}, the output of \code{\link{mBenchmarkMotifs}}}

\item{type}{benchmark (test if known motifs were observed) in query proteins or in all proteins the sequences of which were used in QSLIMFinder search}
}
\value{
benchmarkMotifsROC: list of ROCR objects (prediction, performance, auc.performance) used to plot ROC curve

mBenchmarkMotifsROC: list of lists of ROCR objects (prediction, performance, auc.performance) used to plot ROC curve
}
\description{
Plot ROC or precision-recall curves for motifs search benchmark
}
\seealso{
\code{\link{benchmarkMotifs}}

\code{\link{benchmarkMotifs}}
}
\author{
Vitalii Kleshchevnikov

Vitalii Kleshchevnikov
}
