% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgelist2degree.R
\name{edgelist2degree}
\alias{edgelist2degree}
\title{find the degree for each protein given molecular interaction data containing unique pair_id}
\usage{
edgelist2degree(mitab, sep = "\\\\|")
}
\arguments{
\item{mitab}{data.table containing molecular interaction data, including pair_id}

\item{sep}{how are individual interactors in the pair_id reparated? default: "\\|"}
}
\value{
sorted data.table containing degree for each of the nodes in the input, details: \code{mitab}
}
\description{
find the degree for each protein given molecular interaction data containing unique pair_id
}
\details{
unique pair_id is the unique identifier of the undirected interaction: ordered alphabetically and concatenated interacting molecule IDs \code{\link{cleanMITAB}}
}
\examples{
edgelist = data.table(pair_id = paste0(rep(c("P", "Q", "R"), each = 30),
                      "|",
                      sample(rep(c("Z", "X", "F"), each = 30)), 1:90))
edgelist2degree(edgelist)
}
\author{
Vitalii Kleshchevnikov
}
