% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgelist2degreeFreq.R
\name{edgelist2degreeFreq}
\alias{edgelist2degreeFreq}
\title{Find the degree distribution given edgelist (pair_id) and a list of proteins which degrees matter}
\usage{
edgelist2degreeFreq(mitab, prots, sep)
}
\arguments{
\item{mitab}{data.table containing molecular interaction data, including pair_id}

\item{prots}{list of relevant proteins, degree distribution is based on the prevalence of specific degrees in this list, all proteins by default}

\item{sep}{ID separator in pair_id, passed to \code{\link{edgelist2degree}}}
}
\description{
Find the degree distribution given edgelist (pair_id) and a list of proteins which degrees matter
}
\details{
unique pair_id is the unique identifier of the undirected interaction: ordered alphabetically and concatenated interacting molecule IDs \code{\link{cleanMITAB}}
}
\author{
Vitalii Kleshchevnikov
}
