% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foldEnrichmentDist.R
\name{foldEnrichmentDist}
\alias{foldEnrichmentDist}
\title{find fold_enrichment distribution (in any domain or other feature) for each protein in the network}
\usage{
foldEnrichmentDist(net, protein_annot, N = 1000, cores = NULL, seed = 1,
  frequency = T)
}
\arguments{
\item{net}{data.table that specifies the PPI network:
1. first column specifies a protein ("IDs_interactor_viral")
2. second column specifies it's interacting partner which annotations are of interest ("IDs_interactor_human")
3. third column specifies the degree of a protein in the first column ("IDs_interactor_viral_degree")}

\item{protein_annot}{data.table that specifies annotations for a protein in the second column of \code{net}
1. first column specifies a protein ("IDs_interactor_human")
2. second column specifies features of that protein ("IDs_domain_human")
3. third column specifies background frequency of those features ("domain_frequency)}

\item{N}{number of times to run permutation of PPI network}

\item{cores}{specify how many cores to use for parallel processing, default (NULL) is to detect all cores on the machine and use all minus one. When using LSF cluster you must specify the number of cores to use because \code{\link[BiocGenerics]{detectCores}} doen't know how much cores you have requested from LSF (with bsub -n) and detects all cores on the actual physical node.}

\item{seed}{seed for RNG for reproducible sampling}

\item{frequency}{fold enrichment or frequency in a set (if TRUE - frequency), also passed to \code{\link[MItools]{foldEnrichment}}}
}
\value{
data.table containing fold enrichment for each domain (or other feature) - protein pair (2 columns: IDs_interactor_viral and sampled_fold_enrichment or sampled_domain_frequency_per_set)
}
\description{
\code{foldEnrichmentDist} finds fold_enrichment distribution (in any domain or other feature) for each protein in the first column of net. For more details: \code{\link[MItools]{foldEnrichment}}
}
\author{
Vitalii Kleshchevnikov
}
