% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foldEnrichmentPval.R
\name{foldEnrichmentPval}
\alias{foldEnrichmentPval}
\title{Find pvalue for each protein-feature pair in the network}
\usage{
foldEnrichmentPval(fold_enrichment_dist, data, cores = NULL)
plotFoldEnrichmentDist(proteinID, fold_enrichment_dist, data, main = NULL, text_lab = T)
}
\arguments{
\item{fold_enrichment_dist}{data.table containing fold enrichment for each domain (or other feature) - protein pair (2 columns: IDs_interactor_viral and sampled_fold_enrichment), generated by \code{\link[MItools]{foldEnrichmentDist}}}

\item{data}{data.table containing foldEnrichment value for each protein-feature pair (4 columns: IDs_interactor_viral, IDs_interactor_human, IDs_domain_human, fold_enrichment)}

\item{cores}{specify how many cores to use for parallel processing, default (NULL) is to detect all cores on the machine and use all minus one. When using LSF cluster you must specify the number of cores to use because \code{\link[BiocGenerics]{detectCores}} doen't know how much cores you have requested from LSF (with bsub -n) and detects all cores on the actual physical node.}

\item{frequency}{fold enrichment or frequency in a set (if TRUE - frequency)}

\item{proteinID}{character specifying proteinID for which to plot fold Enrichment distribution}

\item{main}{title for the plot}

\item{text_lab}{logical: label domains (features) on the plot, or not}
}
\value{
data.table containing pvalue for each protein-feature pair in the network (4 columns: proteinID, IDs_interactor_human, featureID, foldEnrichment and Pval)
}
\description{
\code{foldEnrichmentPval} calculates how often each fold Enrichment value for the protein-feature pair is as high or higher than the foldEnrichment value for the protein - any (random) feature pairs (obtained from permutations using \code{\link[MItools]{foldEnrichmentDist}}).

\code{plotFoldEnrichmentDist} allows to visualise sampled fold Enrichment distribution and the real fold Enrichment for the protein-feature pair for specific protein
}
\author{
Vitalii Kleshchevnikov
}
