% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humanViralDegree.R
\name{loadHumanViralPPI}
\alias{loadHumanViralPPI}
\title{load human-human, human-viral and viral-viral PPI data from IntActFTP}
\usage{
loadHumanViralPPI(directory = "./data_files/")
}
\arguments{
\item{where}{to find / keep PPI data (arg for \code{\link[MItools]{fullInteractome}}, \code{\link[MItools]{interSpeciesInteractome}}))}
}
\description{
use \code{\link[MItools]{interSpeciesInteractome}} and \code{\link[MItools]{fullInteractome}} to retrieve and load human-human, human-viral and viral-viral PPI data from IntActFTP

to be used with \code{\link[MItools]{humanViralDegree}}
}
\author{
Vitalii Kleshchevnikov
}
