% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadIntActFTP.R
\name{loadIntActFTP}
\alias{loadIntActFTP}
\title{load all data from IntAct ftp}
\usage{
loadIntActFTP(dir, release = NULL)
}
\arguments{
\item{dir}{directory where to save/look for the local copy}

\item{release}{which locally saved IntAct release to load (the default is to load the latest and read it into R)}
}
\value{
\code{loadIntActFTP} saves intact.txt to a file, returns the object of class RAW_MItab27
}
\description{
\code{loadIntActFTP} loads all data stored on \link{ftp://ftp.ebi.ac.uk/pub/databases/intact/current/psimitab/} in intact.txt file. Details: \link{http://www.ebi.ac.uk/intact/downloads}. This file contains the data from the following databases "IntAct", "MINT", "DIP", "bhf-ucl", "MPIDB", "MatrixDB", "HPIDb","I2D-IMEx","InnateDB-IMEx", "MolCon", "UniProt", "MBInfo"
}
\examples{
loadIntActFTP("./", release = NULL)
}
\author{
Vitalii Kleshchevnikov

Vitalii Kleshchevnikov
}
