% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QSLIMFinderCommand.R
\name{QSLIMFinderCommand}
\alias{QSLIMFinderCommand}
\alias{mQSLIMFinderCommand}
\alias{groupQSLIMFinderCommand}
\title{Generate a QSLIMFinder command}
\usage{
QSLIMFinderCommand(file_list, i = 1,
  slimpath = "../software/slimsuite/tools/",
  blast = "../software/ncbi_blast_2.6.0/bin/",
  iupred = "../software/iupred/iupred",
  options = "dismask=T consmask=F cloudfix=F probcut=0.1",
  LSF_cluster_par = "bsub -n 1 -q research-rh7 -M 100 -R \\"rusage[mem=100]\\"",
  LSF_project_path = "/hps/nobackup/research/petsalaki/users/vitalii/vitalii/viral_project/",
  analysis_type = c("qslimfinder", "slimfinder")[1])

mQSLIMFinderCommand(file_list, slimpath = "../software/slimsuite/tools/",
  blast = "../software/ncbi_blast_2.6.0/bin/",
  iupred = "../software/iupred/iupred",
  options = "dismask=T consmask=F cloudfix=F probcut=0.5",
  LSF_cluster_mode = F,
  LSF_project_path = "/hps/nobackup/research/petsalaki/users/vitalii/vitalii/viral_project/",
  LSF_cluster_par = "bsub -n 1 -q research-rh7 -M 200 -R \\"rusage[mem=200]\\"",
  log_dir = "./SLIMFinder/log_dir/", write_log = T, recursive = F,
  analysis_type = c("qslimfinder", "slimfinder")[1])

groupQSLIMFinderCommand(commands, InteractionSubsetFASTA_list,
  sh_dir = "/sh_dir/",
  LSF_project_path = "/hps/nobackup/research/petsalaki/users/vitalii/vitalii/viral_project/",
  dataset_name = "SLIMFinder", N_seq = 200, write_log = T)
}
\arguments{
\item{file_list}{data.table containing path to files and directories for QSLIMFinder: fastafile, queryfile, outputdir, outputfile}

\item{i}{integer, which set of files and directories to choose from \code{file_list}}

\item{slimpath}{relative path (from the project folder) to the directory containing qslimfinder.py or slimfinder.py}

\item{blast}{relative path (from the project folder) to /bin/ folder in BLAST package}

\item{iupred}{relative path (from the project folder) to iupred (compiled executable)}

\item{options}{any options from QSLIMFinder}

\item{LSF_cluster_par}{a string that will launch LSF job}

\item{LSF_project_path}{absolute path on LSF cluster}

\item{analysis_type}{character 1L, qslimfinder or slimfinder. slimfinder doesn't need query: identical datasets with the same query will be removed and query files not written)}

\item{LSF_cluster_mode}{logical, if FALSE \code{LSF_cluster_par} and \code{LSF_project_path} are ignored}

\item{log_dir}{character, directory where to write log files (stout and sterr)}

\item{write_log}{whether to save stout and sterr}

\item{commands}{list returned by \code{mQSLIMFinderCommand()} containing: 1. command to set up enviromental variable IUPred_PATH; 2. character vector of bash commands that will lauch QSLIMFinder as a job on LSF cluster; 3, 4, 5 - directories where LSF should write stout and sterr}

\item{InteractionSubsetFASTA_list}{object of class InteractionSubsetFASTA_list containing: FASTA sequences for interacting proteins, molecular interaction data they correspond to. Each element of a list contains input for individual QSLIMFinder run.}

\item{sh_dir}{directory within dataset directory (dataset_name) where to write batch command .sh files}

\item{dataset_name}{character, name of the dataset, such as "SLIMFinder" or "SLIMFinder_Vidal"}

\item{N_seq}{size of the batch. Groups QSLIMFinder jobs until the next job doesn't fit into N_seq, if jobs is larger than N_seq a single job will be written to a batch .sh file.}

\item{LSF_project_path}{full path to the project where dataset directory (dataset_name) is located.}

\item{write_log}{FALSE will not allow runQSLIMFinder to detect crashed jobs}
}
\value{
character vector (1L), bash command that will lauch QSLIMFinder locally or as a job on LSF cluster

list containing: 1. command to set up enviromental variable IUPred_PATH; 2. character vector of bash commands that will lauch QSLIMFinder as a job on LSF cluster; 3, 4, 5 - directories where LSF should write stout and sterr

QSLIMFinderCommand split into batches by N_seq sequences. List containing: 1. command to set up enviromental variable IUPred_PATH; 2. character vector of bash commands that will lauch $SHELL as a job on LSF cluster and run QSLIMFinder commands from batch file; 3, 4, 5 - directories where LSF should write stout and sterr
}
\description{
Generate a QSLIMFinder command
}
\examples{
all_commands = mQSLIMFinderCommand(file_list = forSLIMFinder_file_list,
    slimpath = "../software/cluster/slimsuite/tools/qslimfinder.py",
    blast = "../software/cluster/ncbi_blast_2.6.0/bin/",
    iupred = "../software/cluster/iupred/iupred",
    options = "dismask=T consmask=F cloudfix=T probcut=0.3 iuchdir=T",
    LSF_cluster_mode = T,
    LSF_project_path = "/hps/nobackup/research/petsalaki/users/vitalii/vitalii/viral_project/",
    LSF_cluster_par = "bsub -n 1 -q research-rh7 -M 100 -R \\"rusage[mem=100]\\"")
}
\seealso{
\code{\link{listInteractionSubsetFASTA}}, \code{\link{runQSLIMFinder}}
}
\author{
Vitalii Kleshchevnikov
}
