% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryPSICQUICrlib.R
\name{queryPSICQUICrlib}
\alias{queryPSICQUICrlib}
\title{download PSICQUIC query search result (only IMEx databases) to R library - queryPSICQUIC package - data directory}
\usage{
queryPSICQUICrlib(..., directory = NULL)
}
\arguments{
\item{...}{args passed to \code{\link{queryPSICQUIC}}}
}
\value{
data.table containing the query result
}
\description{
saves PSIQUIC search results to the local R library or to the directory specified
}
\details{
\code{queryPSICQUICrlib} queries IMEx databases using PSICQUIC if the local copy of the query result is not available or if there is a new IntAct release. The functions will still work for non-IMEx databases but it will check the IntAct release date and that is not meaningful for tracking updates of other resources.
}
\examples{
queryPSICQUIC:::downloadPSICQUICrlib(query = "id:P74565 AND detmethod:\\"MI:0018\\"", format = "tab27", database = "imex")
}
\author{
Vitalii Kleshchevnikov
}
