% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterProEntryTypes.R
\name{readInterProGFF3}
\alias{readInterProGFF3}
\alias{addInterProEntryTypes}
\title{Read InterProScan output in the gff format, select sequence features by type}
\usage{
addInterProEntryTypes(InterProScan_result, entry.list_path)
}
\arguments{
\item{InterProScan_result}{GRanges, the output of \code{readInterProGFF3} function}

\item{entry.list_path}{character, file name and directory where to store InterPro entry list}

\item{filename.gz}{character, file name and directory to read InterProScan output in the gff format ("./dir/filename.gz")}

\item{selected_ENTRY_TYPE}{character, select InterPro signature matches from \code{InterProScan_result} using these InterPro Entry Types}
}
\value{
\code{readInterProGFF3}, \code{addInterProEntryTypes} or \code{SubsetByInterProEntryType}: \link[GenomicRanges]{GRanges-class} object containing InterProScan output appended by InterPro Entry Types information, with names metadata and sequence length information imported correctly

\code{getInterProEntryTypes}: data.table containing InterPro Entry List \link{ftp://ftp.ebi.ac.uk/pub/databases/interpro/entry.list}

\code{getInterPro2memberDB}: data.table containing  InterProID to member database ID mapping
}
\description{
Read InterProScan output in the gff format, select sequence features by type
}
\details{
For further details on the InterProScan output format please visit the link \link{https://github.com/ebi-pf-team/interproscan/wiki/OutputFormats}
}
\examples{
# read InterProScan result, download and add InterPro Entry Types information, extract from relevant columns and add names metadata and sequence length information
InterProScan_result = readInterProGFF3("./processed_data_files/all_human_viral_protein_domains.gff3.gz")
InterProScan_result = addInterProEntryTypes(InterProScan_result, "./data_files/entry.list")
# create a subset that contains "Domain", "Active_site", "Binding_site", "Conserved_site", "PTM" signatures
InterProScan_domains = SubsetByInterProEntryType(InterProScan_result)
}
\author{
Vitalii Kleshchevnikov
}
