% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanMITAB27.R
\name{reorderMITAB27}
\alias{reorderMITAB27}
\title{\code{reorderMITAB27} reorders interacting molecules in a pair (and all the corresponding columns) according to order provided in IDs_A_order and IDs_B_order columns (latter are deleted)}
\usage{
reorderMITAB27(mitab)
}
\arguments{
\item{mitab}{data.table containing molecular interaction data in MITAB 2.7 format}
}
\description{
function called internally by \code{\link{cleanMITAB27}} if the format is MITAB2.7
}
\author{
Vitalii Kleshchevnikov
}
