% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runQSLIMFinder.R
\name{runQSLIMFinder}
\alias{runQSLIMFinder}
\alias{modifyMemoryInBsub}
\alias{LSFrunQSLIMFinder}
\title{Run a list of QSLIMFinder jobs}
\usage{
runQSLIMFinder(commands, file_list, max_memory = 5000, onLSF = T,
  recursive = F)

modifyMemoryInBsub(commands, memory)

LSFrunQSLIMFinder(commands)
}
\arguments{
\item{commands}{character vector of commands generated by \code{\link{QSLIMFinderCommand}} that will launch QSLIMFinder jobs}

\item{file_list}{data.table containing path to files and directories for QSLIMFinder: fastafile, queryfile, outputdir, outputfile}

\item{max_memory}{max memory to use for a job}

\item{onLSF}{use LSF cluster to run QSLIMFinder jobs}

\item{recursive}{how to create directories}

\item{memory}{integer, how much memory should the job use}
}
\value{
list: character vector, commands that have crashed even at \code{max_memory}; data.table, corresponding rows of \code{file_list}
}
\description{
Run a list of QSLIMFinder jobs
}
\seealso{
\code{\link{QSLIMFinderCommand}}
}
\author{
Vitalii Kleshchevnikov
}
